% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_ACF.R
\name{acf_weights}
\alias{acf_weights}
\title{Yields the tapering/banding weights from Cai, Ren and Zhou, 2013.}
\usage{
acf_weights(p, n, beta = 0.1, type = 1, k = NULL)
}
\arguments{
\item{p}{The length of the data.}

\item{n}{The number of observations of the data.}

\item{beta}{The beta value (see paper).}

\item{type}{1 for tapering, 2 for banding. Default 1.}

\item{k}{The k value (see paper). If specified, overrides \code{beta}.}
}
\value{
The weight vector.
}
\description{
If both k and beta are specified, k overrides beta.
}
