% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_ACF.R
\name{apply_diag}
\alias{apply_diag}
\title{Applies a summary function to each upper diagonal in X.}
\usage{
apply_diag(X, FUN = mean)
}
\arguments{
\item{X}{The input matrix.}

\item{FUN}{a function which takes in a numerical vector and returns a value.}
}
\value{
The summary function values in a vector, beginning with the one for
the main diagonal and ending with the one for the upper-right corner of X.
}
\description{
Applies a summary function to each upper diagonal in X.
}
