% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{as.matrix2}
\alias{as.matrix2}
\title{\code{as.matrix} and transpose \code{"xifti"}s}
\usage{
as.matrix2(x)
}
\arguments{
\item{x}{The object to coerce to a matrix}
}
\value{
x as a matrix.
}
\description{
A \code{"xifti"} is VxT, whereas \code{clever} and \code{DVARS} accept a
TxV matrix. This function calls \code{as.matrix} and transposes the data
if it is a \code{"xifti"}.
}
\keyword{internal}
