% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choose_PCs.R
\name{choose_PCs.kurtosis}
\alias{choose_PCs.kurtosis}
\title{Selects the components of sufficient kurtosis.}
\usage{
choose_PCs.kurtosis(Comps, kurt_quantile = 0.9, detrend = TRUE, n_sim = 5000)
}
\arguments{
\item{Comps}{A matrix; each column is a component. For PCA, this is the U
matrix. For ICA, this is the M matrix.}

\item{kurt_quantile}{components with kurtosis of at least this quantile are kept.}

\item{detrend}{Should components be detrended before measuring kurtosis? Default is
\code{TRUE}. Recommended if observations represent a time series. Use
\code{FALSE} if they have already been detrended.}

\item{n_sim}{The number of simulation data to use for estimating the sampling
distribution of kurtosis. Only used if a new simulation is performed. (If
\eqn{n<1000} and the quantile is 90%, a pre-computed value is used instead.
If \eqn{n>1000}, the theoretical asymptotic distribution is used instead.)}
}
\value{
The original indices of the components which were retained, in order of
 decreasing kurtosis.
}
\description{
Each component is detrended (this can be disabled) before computing kurtosis. 
 The kurtosis cutoff is the 90% quantile of the sampling distribution of 
 kurtosis for Normal data of the same length as the components; it is
 estimated by simulation or calculated from the theoretical asymptotic
 distribution if the components are long enough.
}
