% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choosePCs.R
\name{choosePCs_kurtosis}
\alias{choosePCs_kurtosis}
\title{Selects the principle components (PCs) of sufficient kurtosis from a SVD.}
\usage{
choosePCs_kurtosis(svd, max_keep = NULL, min_keep = 1, n_sim = 5000)
}
\arguments{
\item{svd}{An SVD decomposition; i.e. a list containing u, d, and v.}

\item{max_keep}{If specified, the total number kept will be at most this
value.}

\item{min_keep}{If specified, the total number kept will be at least this
value.  Default 1.}
}
\value{
The subsetted u matrix with only the chosen columns (PCs).
}
\description{
First, the largest PCs which together explain 90% of the variance are
retained, and smaller components are removed. Each PC is detrended, and the
autocorrelation function of the top 10 PCs is measured. The kurtosis cutoff
is then the 90% quantile of the sampling distribution of kurtosis for
Normal data of the same length and autocorrelation as the PCs; it is
estimated by simulation or calculated from the theoretical asymptotic
distribution if the time series is long enough and autocorrelation is
negligible. Finally, the total number kept is constrained by the
\code{max_keep} and \code{min_keep} arguments.
}
