% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choosePCs.R
\name{choosePCs_variance}
\alias{choosePCs_variance}
\title{Selects the principle components of above-average variance from a SVD.}
\usage{
choosePCs_variance(svd, max_keep = NULL, min_keep = 1)
}
\arguments{
\item{svd}{An SVD decomposition; i.e. a list containing u, d, and v.}

\item{max_keep}{If specified, the total number kept will be at most this
value.}

\item{min_keep}{The total number kept will be at least this
value. The default value is 1.}
}
\value{
A list with the subsetted u matrix with only the chosen columns (PCs),
and the original indices of the PCs which were retained.
}
\description{
PCs with above-average variance are retained, but the total number kept is
constrained by the \code{max_keep} and \code{min_keep} arguments.
}
