% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clever.R
\name{clever}
\alias{clever}
\title{Calculates PCA leverage or robust distance and identifies outliers.}
\usage{
clever(
  x,
  choosePCs = c("kurtosis", "variance"),
  kurt_quantile_cut = 0.9,
  kurt_detrend = TRUE,
  method = c("leverage", "robdist_subset", "robdist"),
  id_out = TRUE
)
}
\arguments{
\item{x}{A wide (obs x vars) data matrix of values.}

\item{choosePCs}{The method to use for choosing which PCs to retain. Default is kurtosis.}

\item{kurt_quantile_cut}{The cutoff quantile to use if \code{choosePCs} is \code{kurtosis}. Default is .9.}

\item{kurt_detrend}{If \code{choosePCs} is \code{kurtosis}, should PCs be detrended before measuring kurtosis?
Default is TRUE. Recommended if observations represent a time series.}

\item{method}{The method to use to measure outlyingness. Default is leverage.}

\item{id_out}{If TRUE (default), will label outliers based on leverage or distance.}
}
\value{
A clever object, i.e. a list with components
\describe{
	\item{params}{A list with the \code{choosePCs}, \code{kurt_quantile_cut}, and \code{method} arguments used.}
 \item{PCs}{The PCs selected by \code{choosePCs}, i.e. the subsetted U matrix from the SVD.}
 \item{leverage}{The leverage of each observation. NULL if \code{method} is not PCA leverage.}
 \item{robdist}{The robust distance of each observation. NULL if \code{method} is not robust distance (subset).}
 \item{inMCD}{Whether each observation is within the MCD subset. NULL if \code{method} is not robust distance (subset).}
 \item{outliers}{An n x 3 data.frame indicating if each observation is an outlier at each of the three levels.}
}
}
\description{
Calculates PCA leverage or robust distance and identifies outliers.
}
\examples{
n_voxels = 1e4
n_timepoints = 100
x = matrix(rnorm(n_timepoints*n_voxels), ncol = n_voxels)

lev = clever(x)
}
