% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clever_convert.R
\name{clever_from_multi}
\alias{clever_from_multi}
\title{Convert \code{"clever_multi"} to \code{"clever"}}
\usage{
clever_from_multi(clev)
}
\arguments{
\item{clev}{The \code{"clever_multi"} object to convert}
}
\value{
The resulting \code{"clever"} object
}
\description{
Convert \code{"clever_multi"} to \code{"clever"}
}
\keyword{internal}
