% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clever_multi.R
\name{clever_multi}
\alias{clever_multi}
\title{Compare multiple leverage measures with \code{clever_multi}}
\usage{
clever_multi(
  X,
  projection = "PCA_kurt",
  nuisance = cbind(1, dct_bases(nrow(X), 4)),
  center = TRUE,
  scale = TRUE,
  kurt_quantile = 0.95,
  PCATF_kwargs = NULL,
  get_dirs = FALSE,
  full_PCA = FALSE,
  get_outliers = TRUE,
  cutoff = 4,
  verbose = FALSE
)
}
\arguments{
\item{projection}{Leverage works by projecting the data onto directions likely to 
 contain outlier information. Choose at least one of the following:

 \describe{
   \item{\code{"PCA"}}{PCA using the top \eqn{k} PCs.}
   \item{\code{"PCA_kurt"}}{PCA using the high-kurtosis PCs among the top \eqn{k}.}
   \item{\code{"PCA2"}}{PCA using the top \eqn{k2} PCs.}
   \item{\code{"PCA2_kurt"}}{PCA using the high-kurtosis PCs among the top \eqn{k2}.}
   \item{\code{"PCATF"}}{PCATF using the top \eqn{k} trend-filtered PCs.}
   \item{\code{"PCATF_kurt"}}{PCATF using the high-kurtosis trend-filtered PCs among the top \eqn{k}.}
   \item{\code{"PCATF2"}}{PCATF using the top \eqn{k2} trend-filtered PCs.}
   \item{\code{"PCATF2_kurt"}}{PCATF using the high-kurtosis trend-filtered PCs among the top \eqn{k2}.}
   \item{\code{"ICA"}}{ICA using the top \eqn{k} ICs.}
   \item{\code{"ICA_kurt"}}{ICA using the high-kurtosis ICs among the top \eqn{k}.}
   \item{\code{"ICA2"}}{ICA using the top \eqn{k2} ICs.}
   \item{\code{"ICA2_kurt"}}{ICA using the high-kurtosis ICs among the top \eqn{k2}.}
 }

 where \eqn{k} is the number of PCs selected by PESEL, and \eqn{k2} is the number
 of PCs with above-average variance.
 
 Use \code{"all"} to use all projection methods. Default: \code{"PCA_kurt"}.}
}
\value{
A \code{"clever_multi"} object, i.e. a list with components
\describe{
 \item{measure}{A \eqn{T \times P} data.frame of numeric leverage values for each of the P projections in \code{projection}.}
 \item{outlier_cutoff}{A \eqn{1 \times P} data.frame of numeric outlier cutoff values for each projection (\code{cutoff} times the median leverage).}
 \item{outlier_flag}{A \eqn{T \times P} data.frame of logical values where \code{TRUE} indicates suspected outlier presence.}
 \item{mask}{
   A length \eqn{P} numeric vector corresponding to the data locations in \code{X}. Each value indicates whether the location was masked:
   \describe{
     \item{1}{The data location was not masked out.}
     \item{-1}{The data location was masked out, because it had at least one \code{NA} or \code{NaN} value.}
     \item{-2}{The data location was masked out, because it was constant.}
   }
 }
 \item{PCA}{
   This will be a list with components:
   \describe{
     \item{U}{The \eqn{T \times Q} PC score matrix.}
     \item{D}{The standard deviation of each PC.}
     \item{V}{The \eqn{P \times Q} PC directions matrix. Included only if \code{get_dirs}.}
     \item{highkurt}{The length \code{Q} logical vector indicating scores of high kurtosis.}
     \item{nPCs_PESEL}{The number of PCs selected by PESEL.}
     \item{nPCs_avgvar}{The number of above-average variance PCs.}
   }
   where \code{Q} is the number of PCs selected by PESEL or of above-average variance (or the greater of the two if both were used). 
   If PCA was not used, all entries except \code{nPCs_PESEL} and/or \code{nPCs_avgvar} will not be included, depending on which
   method(s) was used to select the number of PCs.
 }
 \item{PCATF}{
   If PCATF was used, this will be a list with components:
   \describe{
     \item{U}{The \eqn{T \times Q} PC score matrix.}
     \item{D}{The standard deviation of each PC.}
     \item{V}{The \eqn{P \times Q} PC directions matrix. Included only if \code{get_dirs}}
   }
 }
 \item{ICA}{
   If ICA was used, this will be a list with components:
   \describe{
     \item{S}{The \eqn{P \times Q} source signals matrix. Included only if \code{get_dirs}} 
     \item{M}{The \eqn{T \times Q} mixing matrix.}
     \item{highkurt}{The length \code{Q} logical vector indicating mixing scores of high kurtosis.}
   }
 }
}
}
\description{
Calculates leverage to identify outliers in high-dimensional data. 
 Can compute multiple kinds of leverage at once.
}
\examples{
n_voxels = 1e4
n_timepoints = 100
X = matrix(rnorm(n_timepoints*n_voxels), ncol = n_voxels)

clev = clever:::clever_multi(X)
}
\keyword{internal}
