% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rox_args_docs.R
\name{clever_order_of_operations}
\alias{clever_order_of_operations}
\title{Order of operations in \code{clever}}
\description{
Order of operations in \code{clever}
}
\section{Order of operations}{

The input data undergo a nuisance regression to remove low frequency trends
(see the \code{DCT} argument to control or disable this behavior) and any
other nuisance signals (see the \code{nuisance_too} argument to specify them).
(If \code{DCT==0} and \code{is.null(nuisance_too)}, no nuisance regression
is performed.) Then the data are centered and scaled (see the \code{center}
and \code{scale} arguments to control or disable this behavior). Leverage
is computed afterward.
}

\keyword{internal}
