% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clever_convert.R
\name{clever_to_multi}
\alias{clever_to_multi}
\title{Convert \code{"clever"} to \code{"clever_multi"}}
\usage{
clever_to_multi(clev)
}
\arguments{
\item{clev}{The code{"clever"} object to convert}
}
\value{
The resulting \code{"clever_multi"} object
}
\description{
Not perfect, since some information has been lost.
}
\keyword{internal}
