% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompCor.R
\name{CompCor}
\alias{CompCor}
\title{CompCor}
\usage{
CompCor(
  X,
  ROI_data = "infer",
  ROI_noise = NULL,
  noise_nPC = 5,
  noise_erosion = NULL,
  center = TRUE,
  scale = TRUE,
  nuisance = NULL
)
}
\arguments{
\item{center, scale}{Center the columns of the noise ROI data by their medians, 
and scale by their MADs? Default: \code{TRUE} for both. Note that this argument
affects the noise ROI data and not the data that is being cleaned with aCompCor.
Centering and scaling of the data being cleaned can be done after this function call.}

\item{nuisance}{Nuisance signals to regress from each data column in addition
to the noise ROI PCs. Should be a \eqn{T \times N} numeric matrix where 
\eqn{N} represents the number of nuisance signals. To not perform any nuisance
regression set this argument to \code{NULL}, \code{0}, or \code{FALSE}.
Default: \code{NULL}.}
}
\value{
A list with entries \code{"data"} and \code{"noise"}

 The entry \code{"data"} will be a \code{V x T} matrix where each row is a 
 data voxel (if it was originally an array, the voxels will be in spatial
 order) time series with each noise PC regressed from it. Otherwise, this 
 entry will be \code{NULL}.

 The entry \code{"noise"} is a list of \code{T} \eqn{x} \code{noise_nPC} PC 
 scores, one for each \code{ROI_noise}.
}
\description{
The aCompCor algorithm (Behzadi et. al., 2007) 10.1016/j.neuroimage.2007.04.042
}
\details{
First, the principal components (PCs) of each noise ROI are calculated. For each ROI,
 voxels are centered and scaled (can be disabled with the arguments 
 \code{center} and \code{scale}), and then the PCs are calculated via the
 singular value decomposition. 

Next, aCompCor is performed to remove the shared variation between the noise ROI
 PCs and each location in the data. This is accomplished by a nuisance regression
 using a design matrix with the noise ROI PCs, any additional regressors specified
 by \code{nuisance}, and an intercept term. (To detrend the data and perform aCompCor
 in the same regression, \code{nuisance} can be set to DCT bases obtained with
 the function \code{\link{dct_bases}}.)
}
