% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompCor.R
\name{CompCor.regress}
\alias{CompCor.regress}
\title{CompCor: regress}
\usage{
CompCor.regress(X, noise_comps)
}
\arguments{
\item{X}{The data}

\item{noise_comps}{The noise components}
}
\value{
The data with the noise components regressed from it.
}
\description{
CompCor: regress
}
\keyword{internal}
