% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompCor_HCP.R
\name{CompCor_HCP}
\alias{CompCor_HCP}
\title{\code{clever} with CompCor for HCP data}
\usage{
CompCor_HCP(
  cii,
  brainstructures = c("left", "right"),
  nii,
  nii_labels,
  ROI_noise = c("wm_cort", "csf"),
  noise_nPC = 5,
  noise_erosion = NULL,
  ...
)
}
\arguments{
\item{cii}{\code{"xifti"} (or file path to the CIFTI) from which the noise
ROI components will be regressed. In the HCP, the corresponding file is e.g.
"Results/rfMRI_REST1_LR/rfMRI_REST1_LR_Atlas_MSMAll.dtseries.nii"}

\item{brainstructures}{Choose among "left", "right", and "subcortical".
Default: \code{c("left", "right")} (cortical data only)}

\item{nii}{\eqn{I \times J \times \K \times T} 
NIFTI object or array (or file path to the NIFTI) which contains
whole-brain data, including the noise ROIs. In the HCP, the corresponding
file is e.g. "Results/rfMRI_REST1_LR/rfMRI_REST1_LR.nii.gz"}

\item{nii_labels}{\eqn{I \times J \times K}
NIFTI object or array (or file path to the NIFTI) which
contains the corresponding labels to each voxel in \code{nii}. Values should
be according to this table: 
https://surfer.nmr.mgh.harvard.edu/fswiki/FsTutorial/AnatomicalROI/FreeSurferColorLUT .
In the HCP, the corresponding file is "ROIs/Atlas_wmparc.2.nii.gz".}

\item{ROI_noise}{A list of numeric vectors. Each entry should represent labels
 in \code{nii_labels} belonging to a single noise ROI, named by that entry's 
 name. Or, this can be a character vector of at least one of the following: 
 \code{"wm_cort"} (cortical white matter), \code{"wm_cblm"} (cerebellar white
 matter), \code{"csf"} (cerebrospinal fluid). In the latter case, these labels
 will be used:

 \describe{
   \item{wm_cort}{c(3000:4035, 5001, 5002)}
   \item{wm_cblm}{c(7, 46)}
   \item{csf}{c(4, 5, 14, 15, 24, 31, 43, 44, 63, 250, 251, 252, 253, 254, 255))}
 }

 These default ROIs are based on this forum post: 
 https://www.mail-archive.com/hcp-users@humanconnectome.org/msg00931.html

 Default: \code{c("wm_cort", "csf")}}

\item{noise_nPC, noise_erosion}{See \code{\link{clever}}.}

\item{...}{Additional arguments to \code{\link{clever}}.}
}
\description{
Wrapper to \code{\link{clever}} for computing CompCor (and other outlyingness
 measures) on HCP data. The whole-brain NIFTI is used to obtain the noise
 ROIs, which are regressed from the greyordinate data in the CIFTI.
}
