% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DVARS.R
\name{compute_DVARS}
\alias{compute_DVARS}
\title{Computes the DSE decomposition and DVARS-related statistics.
Citation: Insight and inference for DVARS (Afyouni and Nichols, 2018)}
\usage{
compute_DVARS(X, normalize = FALSE, norm_I = 100, verbose = FALSE)
}
\arguments{
\item{X}{a T x N numeric matrix representing an fMRI run.}

\item{normalize}{}

\item{norm_I}{}

\item{verbose}{Should occasional updates be printed? Default is \code{FALSE}.}
}
\description{
Differences from implementation at github.com/asoroosh/DVARS:
* The matrix is transposed.
* We center and scale the matrix differently (see clever/scale_med)
* We set all zero-variance voxels to zero during centering & scaling. This means
   that when we remove constant 0 or NA voxels, constant non-zero voxels are also 
   removed.
* We use a tolerance of \eqn{1e-8} to detect non-zero voxels.
}
