% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Dat1}
\alias{Dat1}
\title{Example fMRI Time Series for Subject 1}
\format{A matrix of 193 time points by 4675 voxels}
\usage{
Dat1
}
\description{
A masked slice of an fMRI time series, turned into a matrix.
}
\keyword{datasets}
