% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Dat2}
\alias{Dat2}
\title{Example fMRI Time Series for Subject 2}
\format{A matrix of 145 time points by 4679 voxels}
\usage{
Dat2
}
\description{
A masked slice of an fMRI time series, turned into a matrix
}
\keyword{datasets}
