% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dct_bases}
\alias{dct_bases}
\title{Get the cosine bases for DCT}
\usage{
dct_bases(T_, n)
}
\arguments{
\item{T_}{Length of timeseries}

\item{n}{Number of cosine bases}
}
\value{
Matrix with cosine bases along columns
}
\description{
Get the cosine bases for DCT
}
