% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dct_mat}
\alias{dct_mat}
\title{Multiple timecourse DCT}
\usage{
dct_mat(
  x,
  add_mean = TRUE,
  n_bases = 4,
  highpass_freq = NULL,
  TR = NULL,
  residual = TRUE
)
}
\arguments{
\item{x}{A matrix. (Vectors will be coerced to a matrix with one column.)}

\item{add_mean}{Add constant column to design matrix to de-mean the data?
Default: \code{TRUE}.}

\item{n_bases}{Number of cosine bases. Default: \code{4}.}

\item{highpass_freq, TR}{Not implemented yet; use \code{n_bases}}

\item{residual}{\code{TRUE} (default) to get the residuals after regressing
the cosine bases (detrend using DCT), or \code{FALSE} to get the estimate
of that regression (trendline using DCT).}
}
\value{
The estimated detrended timeseries, or trend, based on the DCT.
}
\description{
Detrends, or estimates the trend, of each column in a matrix using the
 discrete cosine transform (DCT).
}
