% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SHASH.R
\name{emprule_rob}
\alias{emprule_rob}
\title{Robust empirical rule}
\usage{
emprule_rob(x, thr = 4)
}
\arguments{
\item{x}{The data}

\item{thr}{MAD threshold}
}
\value{
Logical vector indicating whether each element in \code{x} is an
outlier (\code{TRUE} if an outlier).
}
\description{
Robust empirical rule outlier detection applicable to approximately Normal data
}
\keyword{internal}
