% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_ACF.R
\name{est_ACF}
\alias{est_ACF}
\title{Estimates the ACF of columns in X.}
\usage{
est_ACF(X, detrend = TRUE, method = "matrixwise",
  cor_meas = "spearman", center_meas = median)
}
\arguments{
\item{X}{A numeric vector representing a length-t time series; or, a
txp matrix representing p length-t time series.}

\item{detrend}{Should the time series be detrended? Default TRUE.}

\item{method}{'matrixwise' will compute the multivariate correlation and
take the center of the diagonals (default). 'vectorwise' will compute the
ACF for each column, and then take the center at each lag. Different
measures of correlation are available for 'matrixwise'; only Pearson
correlation is available for 'vectorwise'.}

\item{cor_meas}{The \code{method} parameter for \code{stats::cor} to
use for ACF estimation with the 'matrixwise' method. Default is 'spearman'.}

\item{center_meas}{How values should be aggregated. Default is
\code{stats::median}.}
}
\value{
The ACF as a numeric vector.
}
\description{
Estimates the ACF of columns in X.
}
