% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_ACF.R
\name{est_trend}
\alias{est_trend}
\title{Estimates the trend of \code{ts} using a cubic spline.}
\usage{
est_trend(ts, n_knots = 5, rm_out = "combined")
}
\arguments{
\item{ts}{A numeric vector.}

\item{n_knots}{The number of knots to use for the spline model.}
}
\value{
The estimated trend.
}
\description{
Estimates the trend of \code{ts} using a cubic spline.
}
