% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FD.R
\name{FD}
\alias{FD}
\title{Framewise Displacement}
\usage{
FD(
  X,
  trans_units = c("mm", "cm", "in"),
  rot_units = c("deg", "rad", "mm", "cm", "in"),
  brain_radius = NULL,
  detrend = FALSE,
  cutoff = 0.3
)
}
\arguments{
\item{X}{An \eqn{N \times 6} matrix in which the first three columns represent the
 translational RPs (\code{trans_units}), and the second three columns represent
 the rotational RPs (\code{rot_units}). If \code{rot_units} measures an angle,
 it will be converted to \code{trans_units} by measuring displacement on a 
 sphere of radius \code{brain_radius} \code{trans_units}.

 Alternatively, this can be the file path to an \eqn{N \times 6} matrix which can be
 read with \code{\link[utils]{read.table}} (fields separated by white-space; no
 header).}

\item{trans_units}{\code{"mm"} for millimeters (default), \code{"cm"} 
for centimeters, or \code{"in"} for inches.}

\item{rot_units}{\code{"deg"} for degrees (default), \code{"rad"} for radians,
or one of the \code{trans_units} options.}

\item{brain_radius}{If \code{rot_units} measures an angle, the rotational RPs
 are transformed to a spatial measurement representing the displacement on a 
 sphere of radius \code{brain_radius} \code{trans_units}.

 If \code{brain_radius} is \code{NULL} (default), its value will be set to 
 (the equivalent of) 50 mm.}

\item{detrend}{Detrend each RP with the DCT before computing FD?
Default: \code{FALSE}. Can be a number of DCT bases to use, or \code{TRUE}
to use 4.}

\item{cutoff}{[TO DO]}
}
\value{
A length \eqn{N} vector of FD values in \code{trans_units}.
}
\description{
Calculates Framewise Displacement (FD)
}
\details{
The FD formula is taken from (Power, 2012):

 \eqn{FD_i = \lvert \Delta x_i \rvert + \lvert \Delta y_i \rvert + \lvert \Delta z_i + \lvert \Delta \alpha_i \rvert + \lvert \Delta \beta_i \rvert + \lvert \Delta \gamma_i \rvert}, 
 where \eqn{i} is the timepoint; \eqn{x}, \eqn{y} and \eqn{z} are the 
 translational realignment parameters (RPs);
 \eqn{\alpha}, \eqn{\beta} and \eqn{\gamma} are the rotational RPs;
 and \eqn{\Delta x_i = x_{i-1} - x_i} (and similarly for the other RPs).

 Citation: 10.1016/j.neuroimage.2011.10.018
}
