% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params_F.R
\name{fit.F}
\alias{fit.F}
\title{Estimates the parameters of the F distribution of MCD distances}
\usage{
fit.F(Q, n, h)
}
\arguments{
\item{Q}{number of variables in dataset used to compute MCD distances}

\item{n}{total number of observations}

\item{h}{number of observations included in estimation of MCD center and scale}
}
\description{
This estimates the parameters c and m required to determine the distribution
of robust MCD distances as derived by Hardin and Rocke (2005), The Distribution
of Robust Distances.
}
