% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsl_bptf.R
\name{fsl_bptf}
\alias{fsl_bptf}
\title{bptf function from FSL}
\usage{
fsl_bptf(orig_data, HP_sigma = 2000)
}
\arguments{
\item{orig_data}{T_ x V_ data matrix whose columns will be detrended}

\item{HP_sigma}{The frequency parameter for the highpass filter}
}
\description{
Copy of highpass filter as implemented by bptf in FSL. References:
 https://cpb-us-w2.wpmucdn.com/sites.udel.edu/dist/7/4542/files/2016/09/fsl_temporal_filt-15sywxn.m
 https://github.com/rordenlab/niimath/blob/master/src/core32.c
}
