% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leverage_images.R
\name{get_leverage_images}
\alias{get_leverage_images}
\title{Calculate the leverage images for each outlier that meets the
 \code{outlier_level} threshold, with 3 (default) being the highest/strictest
 and 1 being the lowest.}
\usage{
get_leverage_images(X_svd, timepoints, const_mask = NULL)
}
\arguments{
\item{X_svd}{A singular value decomposition (list with u, v, and d).}

\item{timepoints}{The times for which to compute leverage images (rows of U).}

\item{const_mask}{Mask for the voxels that were removed.}
}
\value{
A list of three: the mean leverage image for each outlier meeting
 the thresold, the top leverage image for each outlier, and the indices of
 the top leverage images.
}
\description{
Calculate the leverage images for each outlier that meets the
 \code{outlier_level} threshold, with 3 (default) being the highest/strictest
 and 1 being the lowest.
}
