% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_mad}
\alias{get_mad}
\title{Computes the median absolute deviation for columns in a matrix.}
\usage{
get_mad(mat)
}
\arguments{
\item{mat}{A numerical matrix.}
}
\value{
A vector with the median absolute deviations of each column in the input matrix.
}
\description{
Computes the median absolute deviation for columns in a matrix.
}
