% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompCor.R
\name{hat_mat}
\alias{hat_mat}
\title{Hat matrix}
\usage{
hat_mat(X)
}
\arguments{
\item{X}{Numeric matrix. A column of ones will be added for the intercept.}
}
\value{
The hat matrix, a square matrix with the same number of rows/columns
 as the number of rows in \code{X}.
}
\description{
Get the hat matrix of X after adding a column of ones (for the intercept).
}
\keyword{internal}
