% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_outliers.R
\name{id_out.leverage}
\alias{id_out.leverage}
\title{Identifies outliers based on leverage.}
\usage{
id_out.leverage(leverage)
}
\arguments{
\item{leverage}{A vector of length n with the leverage of each observation.}
}
\value{
A list with components
\describe{
	\item{outliers}{
	An n x 3 data.frame indicating if each observation is an outlier at the 
	3*median, 4*median, and 5*median levels.
}
\item{cutoffs}{The leverage cutoff values: 3*median, 4*median, and 5*median.}
}
}
\description{
Observations with leverage greater than 3, 4, and 5 times the median are labeled as outliers.
}
