% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_outliers.R
\name{id_out.robdist}
\alias{id_out.robdist}
\title{Identifies outliers based on robust distance.}
\usage{
id_out.robdist(distance, inMCD, Fparam)
}
\arguments{
\item{distance}{A vector of length n with the robust distance estimate of each observation.}

\item{inMCD}{A vector of length n indicating if each observation is within the MCD subset.}

\item{Fparam}{The estimated parameters of the F distribution of MCD distances.}
}
\value{
A list with components
\describe{
	\item{outliers}{
	An n x 3 data.frame indicating if each observation is an outlier at the 
	1e-2, 1e-3, and 1e-4 quantile levels.
}
\item{cutoffs}{The robust distance cutoff values: the 1e-2, 1e-3, and 1e-4th quantiles}
}
}
\description{
Observations whose robust distance lies in the top 1e-2, 1e-3, and 1e-4
 quantiles of the estimated F distribution are labeled as outliers.
}
