% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/out_meas.R
\name{id_out_boot_robdist}
\alias{id_out_boot_robdist}
\title{Identify outliers based on bootstrap robust distance.}
\usage{
id_out_boot_robdist(
  dep_data,
  R_true,
  boot_sample = 1000,
  quantile_cutoff = 0.9999
)
}
\arguments{
\item{dep_data}{N x P Dependent data}

\item{R_true}{N x N correlation matrix}

\item{boot_sample}{Number of bootstrap samples to compute}

\item{quantile_cutoff}{Cutoff quantile for outliers}
}
\value{
List of "cut" (P cutoff values, one for each observation) and "flag" (P indicators of outlyingness)
}
\description{
Identify outliers based on bootstrap robust distance.
}
