% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/out_meas.R
\name{induc_indep}
\alias{induc_indep}
\title{Induce independence}
\usage{
induc_indep(dep_data, R_true)
}
\arguments{
\item{dep_data}{N x P Dependent data}

\item{R_true}{N x N correlation matrix}
}
\value{
List of "indep_data" (independent data) and "R_sqrt" (square root of R_true)
}
\description{
Induce independence
}
