% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_ACF.R
\name{k_choice}
\alias{k_choice}
\title{k from Cai, Ren, and Zhou, 2013}
\usage{
k_choice(beta, n, p)
}
\arguments{
\item{beta}{The beta value (see paper).}

\item{n}{The number of observations of the data.}

\item{p}{The length of the data.}
}
\description{
k from Cai, Ren, and Zhou, 2013
}
