% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{logL.lnorm}
\alias{logL.lnorm}
\title{Computes the log likelihood of a sample of values from a log normal distribution.}
\usage{
logL.lnorm(par, vals, cutoff)
}
\arguments{
\item{par}{A vector of length two which contains the degrees of freedom values.}

\item{vals}{A vector of values for which log likelihood will be calculated.}

\item{cutoff}{Values greater than the cutoff are removed before log likelihood calculation.}
}
\value{
A scalar which represents the log likelihood.
}
\description{
Computes the log likelihood of a sample of values from a log normal distribution.
}
