% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{Matrix_to_VolumeTimeSeries}
\alias{Matrix_to_VolumeTimeSeries}
\title{Converts a vectorized matrix back to a volume time series.}
\usage{
Matrix_to_VolumeTimeSeries(mat, mask, sliced.dim = NA)
}
\description{
Converts a vectorized matrix back to a volume time series.
}
