% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nuisance_reg.R
\name{nuisance_reg}
\alias{nuisance_reg}
\title{Nuisance regression}
\usage{
nuisance_reg(dat, nuis = NULL, intercept = TRUE)
}
\arguments{
\item{dat}{\eqn{N \times P} matrix of N measurements/timepoints, P data variables}

\item{nuis}{\eqn{N \times Q} matrix of N measurements/timepoints, Q nuisance regressors.
If \code{NULL}, only regress intercept (if \code{intercept==TRUE}. otherwise, return the original data).}

\item{intercept}{Add intercept column (constant) to \code{nuis}? Default: \code{TRUE}.}
}
\value{
\code{dat} with \code{nuis} regressed from it
}
\description{
Regresses matrix of nuisance variables from data matrix
}
