% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompCor.R
\name{nuisance_regression}
\alias{nuisance_regression}
\title{Nuisance regression}
\usage{
nuisance_regression(X, design)
}
\arguments{
\item{X}{The TxV or VxT data.}

\item{design}{The TxQ matrix of nuisance regressors}
}
\value{
The data after nuisance regression
}
\description{
Performs nuisance regression. The data and design matrix must both be
 centered, or an intercept must be included in the design matrix!
}
\keyword{internal}
