% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/out_measures.R
\name{out_measures.leverage}
\alias{out_measures.leverage}
\title{Computes PCA leverage.}
\usage{
out_measures.leverage(U, median_cutoff = NA)
}
\arguments{
\item{U}{An n x Q matrix of PC scores.}

\item{median_cutoff}{The outlier cutoff, in multiples of the median leverage.
Default: \code{NA} (do not compute outliers).}
}
\value{
A list with entries \code{"meas"} (the leverage values), 
 \code{"cut"} (the leverage cutoff value) and 
 \code{"flag"} (logical vector indicating the outliers). If 
 \code{!is.null(median_cutoff)}, all entries except \code{"meas"} are omitted.
}
\description{
Computes the leverage of each observation, in the PC score matrix (U).
 Optionally can identify the outliers.
}
