% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robdist.R
\name{out_measures.robdist}
\alias{out_measures.robdist}
\title{Computes MCD distances.}
\usage{
out_measures.robdist(Comps, quantile_cutoff = NULL)
}
\arguments{
\item{Comps}{An n x Q matrix of PC scores.}

\item{quantile_cutoff}{The F-distribution quantile cutoff. Default:
\code{NULL} (do not compute outliers).}
}
\value{
A list with entries
\describe{
\item{"meas"}{A vector of length n of with the robust distance estimate
for each observation.}
\item{"info"}{A list with entries "inMCD", "outMCD_scale", and "Fparam"}
\item{"cut"}{The robust distance cutoff value}
\item{"flag"}{Logical vector indicating the outliers}
}

If \code{is.null(quantile_cutoff)} the latter two elements are omitted.
}
\description{
Computes robust minimum covariance determinant (MCD) distances across
the observations (rows). The MCD method selects a subset of h observations
whose covariance matrix has minimum determinant across all subsets of size
h. The MCD distances are Mahalanobis distances using the estimates of
center (mean) and scale (covariance matrix) based on that subset.
}
