% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/out_measures.R
\name{out_measures.robdist_bootstrap}
\alias{out_measures.robdist_bootstrap}
\title{Identify outliers based on bootstrap robust distance.}
\usage{
out_measures.robdist_bootstrap(
  U,
  R_true = NULL,
  boot_sample = 1000,
  quantile_cutoff = NA
)
}
\arguments{
\item{U}{N x P Dependent data}

\item{R_true}{N x N correlation matrix. If \code{NULL} (default), the identity
matrix wil lbe used}

\item{boot_sample}{Number of bootstrap samples to compute}

\item{quantile_cutoff}{The F-distribution quantile cutoff. Default: 
\code{NA} (do not compute outliers).}
}
\value{
List of "meas" (original RD values), "cut" (P cutoff values, one for 
 each observation) and "flag" (P indicators of outlyingness). If \code{is.null(quantile_cutoff)},
 the latter two entries are omitted.
}
\description{
Identify outliers based on bootstrap robust distance.
}
