% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/out_cuts.R
\name{outs.leverage}
\alias{outs.leverage}
\title{Determine the leverage outliers}
\usage{
outs.leverage(lev, median_cutoff = 4)
}
\arguments{
\item{lev}{The leverage values}

\item{median_cutoff}{The cutoff, in multiples of the median of \code{lev}.
Default: \code{4}}
}
\value{
List with entries \code{"cut"} (the leverage cutoff value) and
 \code{"flag"} (logical vector indicating the outliers)
}
\description{
Identify outliers based on leverage values above a multiple of the median.
}
