% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/out_cuts.R
\name{outs.robdist}
\alias{outs.robdist}
\title{Determine the robust distance outliers}
\usage{
outs.robdist(
  rbd,
  Fparam_1,
  Fparam_2,
  inMCD,
  outMCD_scale,
  quantile_cutoff = 0.9999
)
}
\arguments{
\item{rbd}{The robust distance values}

\item{Fparam_1, Fparam_2}{the F distribution degrees of freedom}

\item{inMCD}{Logical vector indicating in-MCD observations}

\item{outMCD_scale}{The scale for out-of-MCD observations}

\item{quantile_cutoff}{The F-distribution quantile cutoff. Default: 
\code{.9999}}
}
\value{
List with entries \code{"cut"} (the robust distance cutoff value for
 scaled out-of-MCD observations) and \code{"flag"} (logical vector 
 indicating the outliers)
}
\description{
Identify outliers based on leverage values above a multiple of the median.
}
