% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCATF.R
\name{PCATF_check_kwargs}
\alias{PCATF_check_kwargs}
\title{Check PCATF args}
\usage{
PCATF_check_kwargs(
  X,
  X.svd,
  solve_directions,
  K,
  lambda,
  niter_max,
  TOL,
  verbose
)
}
\arguments{
\item{X, X.svd, solve_directions, K, lambda, niter_max, TOL, verbose}{See respective
functions.}
}
\value{
\code{NULL}, invisibly.
}
\description{
Check PCATF arguments for both R-core and cpp-core versions
}
\keyword{internal}
