% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_levdist.R
\name{PCleverage}
\alias{PCleverage}
\title{Computes PCA leverage}
\usage{
PCleverage(U)
}
\arguments{
\item{U}{n x Q matrix of PC scores}
}
\value{
a vector of length n with leverage of each observation
}
\description{
Computes the leverage of each observation, based on its PC scores.
}
