% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.clever}
\alias{plot.clever}
\title{Plot \code{"clever"}}
\usage{
\method{plot}{clever}(x, title = NULL, ...)
}
\arguments{
\item{x}{The \code{"clever"} object.}

\item{title}{(Optional) If provided, will add a title to the plot.}

\item{...}{Additional arguments to ggplot: main, sub, xlab, ylab, legend.position}
}
\value{
A ggplot
}
\description{
Plot \code{"clever"}
}
