% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.clever_multi}
\alias{plot.clever_multi}
\title{Plot \code{"clever_multi"}}
\usage{
\method{plot}{clever_multi}(x, measures = "all", title = NULL, ...)
}
\arguments{
\item{x}{The \code{"clever_multi"} object.}

\item{measures}{"all" to plot each measure (default), or a character vector 
of desired measures.}

\item{title}{(Optional) If provided, will add a title to the plot.}

\item{...}{Additional arguments to ggplot: main, sub, xlab, ...}
}
\value{
A ggplot
}
\description{
Plots the outlyingness measures from a \code{"clever_multi"} result.
}
