% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.IDvols_projection_multi}
\alias{plot.IDvols_projection_multi}
\title{Plot a \code{"IDvols_projection_multi"} object}
\usage{
\method{plot}{IDvols_projection_multi}(x, title = NULL, ...)
}
\arguments{
\item{x}{The \code{"IDvols_projection_multi"} object.}

\item{title}{(Optional) If provided, will add a title to the plot.}

\item{...}{Additional arguments to ggplot, e.g. \code{main}, \code{sub},
\code{xlab}, \code{ylab}, \code{legend.position}}
}
\value{
A ggplot
}
\description{
Plot a \code{"IDvols_projection_multi"} object
}
\keyword{internal}
