% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.pscrub_multi}
\alias{plot.pscrub_multi}
\title{Plot \code{"pscrub_multi"}}
\usage{
\method{plot}{pscrub_multi}(x, title = NULL, ...)
}
\arguments{
\item{x}{The \code{"pscrub_multi"} object.}

\item{title}{(Optional) If provided, will add a title to the plot.}

\item{...}{Additional arguments to ggplot, e.g. \code{main}, \code{sub},
\code{xlab}, \code{ylab}, \code{legend.position}}
}
\value{
A ggplot
}
\description{
Plot \code{"pscrub_multi"}
}
\keyword{internal}
