% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.scrub_FD}
\alias{plot.scrub_FD}
\title{Plot a \code{"scrub_FD"} object}
\usage{
\method{plot}{scrub_FD}(x, title = NULL, ...)
}
\arguments{
\item{x}{The \code{"scrub_FD"} object}

\item{title}{(Optional) If provided, will add a title to the plot.}

\item{...}{Additional arguments to ggplot, e.g. \code{main}, \code{sub},
\code{xlab}, \code{ylab}, \code{legend.position}}
}
\value{
A ggplot
}
\description{
Plot a \code{"scrub_FD"} object
}
