% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_scrub_wrapper}
\alias{plot_scrub_wrapper}
\title{Plot scrubbing results}
\usage{
plot_scrub_wrapper(x, title = NULL, ...)
}
\arguments{
\item{x}{The \code{"scrub_*"} object.}

\item{title}{(Optional) If provided, will add a title to the plot.}

\item{...}{Additional arguments to ggplot, e.g. \code{main}, \code{sub},
\code{xlab}, \code{ylab}, \code{legend.position}}
}
\value{
A ggplot
}
\description{
Plot a leverage, DVARS, or FD timeseries from a \code{"scrub_projection"},
\code{"scrub_DVARS"}, or \code{"scrub_FD"} object, respectively. Highlight
volumes flagged for outlier presence.
}
\keyword{internal}
