% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{projection_name}
\alias{projection_name}
\title{Get internal name of this projection}
\usage{
projection_name(projection, PESEL, kurt_quantile)
}
\arguments{
\item{projection, PESEL, kurt_quantile}{See \code{\link{pscrub}}.}
}
\value{
The name of the projection
}
\description{
Get name of projection given PESEL & kurtosis.
}
\keyword{internal}
