% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pscrub_from_multi.R
\name{pscrub_from_multi}
\alias{pscrub_from_multi}
\title{Convert \code{"pscrub_multi"} to \code{"pscrub"}}
\usage{
pscrub_from_multi(psx)
}
\arguments{
\item{psx}{The \code{"pscrub_multi"} object to convert}
}
\value{
The resulting \code{"pscrub"} object
}
\description{
Convert \code{"pscrub_multi"} to \code{"pscrub"}
}
\keyword{internal}
