% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robdist.R
\name{RD_impData}
\alias{RD_impData}
\title{Impute outliers for robust distance}
\usage{
RD_impData(data, univOut)
}
\arguments{
\item{data}{The data}

\item{univOut}{The univariate outliers}

\item{cutoff}{Default: \code{4}}
}
\value{
The data with imputed outliers.
}
\description{
Impute the outliers for robust distance.
}
\keyword{internal}
