% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robdist.R
\name{RD_meas}
\alias{RD_meas}
\title{Robust distance calculation}
\usage{
RD_meas(data, ind_incld, dist = TRUE)
}
\arguments{
\item{data}{The dimension reduced and selected data.}

\item{ind_incld}{Indices of the included subset of \code{data}.}

\item{dist}{Compute the robust distances? Default: \code{TRUE}.}
}
\value{
A list with parameters and optionally the robust distances.
}
\description{
Compute robust distance
}
\keyword{internal}
