% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_ACF.R
\name{reg_ACF}
\alias{reg_ACF}
\title{Regularizes an initial estimate of the ACF, \code{sigma}.}
\usage{
reg_ACF(sigma, method, beta = 0.1, k = NULL, p = NULL)
}
\arguments{
\item{method}{Regularization method: 'tapering' or 'banding' for the
estimators from Cai, Ren and Zhou, 2013; or, 'AR' for an AR(6) model.}

\item{beta}{The beta value for the tapering/banding estimator. See
\code{k_choice()}.}

\item{k}{The k value for the tapering/banding estimator. Overrides \code{beta}.
See \code{k_choice()}.}

\item{p}{The number of time series used to estimate the input sigma. Needed for
tapering or banding estimator.}
}
\value{
The ACF as a numeric vector.
}
\description{
Regularizes an initial estimate of the ACF, \code{sigma}.
}
