% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rob_kurtosis.R
\name{rob_kurtosis}
\alias{rob_kurtosis}
\title{Computes the robust kurtosis measure of a set of values.}
\usage{
rob_kurtosis(x)
}
\arguments{
\item{x}{A vector of values.}
}
\value{
A scalar representing the robust kurtosis.
}
\description{
Computes the robust kurtosis measure of a set of values.
}
\examples{
n = 10
x = rnorm(10)

rob_kurtosis(x)
}
