% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DVARS.R
\name{sd_hIQR}
\alias{sd_hIQR}
\title{Estimate SD robustly using the half IQR}
\usage{
sd_hIQR(x, d = 1)
}
\arguments{
\item{x}{Numeric vector of data to estimate standard deviation for.}

\item{d}{The scalar power transformation parameter. \eqn{w = x^{1/d}} is
computed to obtain \eqn{w \sim N(\mu_w, \sigma_w^2)}.}
}
\value{
Scalar for the robust estimate of standard deviation.
}
\description{
Estimates standard deviation robustly using the half IQR (and power trans.).
Used to measure DVARS in Afyouni and Nichols, 2018.
}
\keyword{internal}
