% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_ACF.R
\name{sim_ts}
\alias{sim_ts}
\title{Simulates a time series with specified length and autocorrelation.}
\usage{
sim_ts(n, ts_length, Sigma = NULL, fit_check = FALSE, fit_tol = 0.2,
  fit_tries = 5, fit_print = FALSE)
}
\arguments{
\item{n}{The number of time series to simulate.}

\item{ts_length}{The length of each time series.}

\item{Sigma}{The correlation matrix. Default is NULL, in which case the
identity matrix is used.}

\item{fit_check}{Should the time series be re-generated if the estimated
correlation deviates significantly from Sigma? Default FALSE.}

\item{fit_tol}{Try again if any (phi_est - phi_true) / phi_true > fit_tol,
with phi being the ACF coefficients. Only applies if fit_tol=TRUE.}

\item{fit_tries}{The maximum number of attempts. Only applies if fit_tol=TRUE.}

\item{fit_print}{Should the estimated ACF coefficients be displayed? (For
debugging.) Default FALSE. Only applies if fit_tol=TRUE.}
}
\value{
A ts_length by n matrix whose columns represent the simulated time
	series.
}
\description{
Simulates a time series with specified length and autocorrelation.
}
