% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_clever.R
\name{summary.clever}
\alias{summary.clever}
\alias{print.summary.clever}
\alias{print.clever}
\title{Summarize a \code{"clever"} object}
\usage{
\method{summary}{clever}(object, ...)

\method{print}{summary.clever}(x, ...)

\method{print}{clever}(x, ...)
}
\arguments{
\item{object}{Object of class \code{"clever"}.}

\item{...}{further arguments passed to or from other methods.}

\item{x}{Object of class \code{"clever"}.}
}
\description{
Summary method for class \code{"clever"}
}
