% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_scrub.R
\name{summary.IDvols_DVARS}
\alias{summary.IDvols_DVARS}
\alias{print.summary.IDvols_DVARS}
\alias{print.IDvols_DVARS}
\title{Summarize a \code{"IDvols_DVARS"} object}
\usage{
\method{summary}{IDvols_DVARS}(object, ...)

\method{print}{summary.IDvols_DVARS}(x, ...)

\method{print}{IDvols_DVARS}(x, ...)
}
\arguments{
\item{object}{Object of class \code{"IDvols_DVARS"}.}

\item{...}{further arguments passed to or from other methods.}

\item{x}{Object of class \code{"IDvols_DVARS"}.}
}
\value{
A plot of the scrubbing results
}
\description{
Summary method for class \code{"IDvols_DVARS"}
}
