% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_scrub.R
\name{summary.IDvols_FD}
\alias{summary.IDvols_FD}
\alias{print.summary.IDvols_FD}
\alias{print.IDvols_FD}
\title{Summarize a \code{"IDvols_FD"} object}
\usage{
\method{summary}{IDvols_FD}(object, ...)

\method{print}{summary.IDvols_FD}(x, ...)

\method{print}{IDvols_FD}(x, ...)
}
\arguments{
\item{object}{Object of class \code{"IDvols_FD"}.}

\item{...}{further arguments passed to or from other methods.}

\item{x}{Object of class \code{"IDvols_FD"}.}
}
\value{
A plot of the scrubbing results
}
\description{
Summary method for class \code{"IDvols_FD"}
}
