% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_scrub.R
\name{summary.IDvols_projection}
\alias{summary.IDvols_projection}
\alias{print.summary.IDvols_projection}
\alias{print.IDvols_projection}
\title{Summarize a \code{"IDvols_projection"} object}
\usage{
\method{summary}{IDvols_projection}(object, ...)

\method{print}{summary.IDvols_projection}(x, ...)

\method{print}{IDvols_projection}(x, ...)
}
\arguments{
\item{object}{Object of class \code{"IDvols_projection"}.}

\item{...}{further arguments passed to or from other methods.}

\item{x}{Object of class \code{"IDvols_projection"}.}
}
\value{
A plot of the scrubbing results
}
\description{
Summary method for class \code{"IDvols_projection"}
}
