% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_clever.R
\name{summary.scrub}
\alias{summary.scrub}
\alias{print.summary.scrub}
\alias{print.scrub}
\title{Summarize a \code{"scrub"} object}
\usage{
\method{summary}{scrub}(object, ...)

\method{print}{summary.scrub}(x, ...)

\method{print}{scrub}(x, ...)
}
\arguments{
\item{object}{Object of class \code{"scrub"}.}

\item{...}{further arguments passed to or from other methods.}

\item{x}{Object of class \code{"scrub"}.}
}
\description{
Summary method for class \code{"scrub"}
}
