% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_scrub.R
\name{summary.scrub_DVARS}
\alias{summary.scrub_DVARS}
\alias{print.summary.scrub_DVARS}
\alias{print.scrub_DVARS}
\title{Summarize a \code{"scrub_DVARS"} object}
\usage{
\method{summary}{scrub_DVARS}(object, ...)

\method{print}{summary.scrub_DVARS}(x, ...)

\method{print}{scrub_DVARS}(x, ...)
}
\arguments{
\item{object}{Object of class \code{"scrub_DVARS"}.}

\item{...}{further arguments passed to or from other methods.}

\item{x}{Object of class \code{"scrub_DVARS"}.}
}
\value{
A plot of the scrubbing results
}
\description{
Summary method for class \code{"scrub_DVARS"}
}
