% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_scrub.R
\name{summary.scrub_FD}
\alias{summary.scrub_FD}
\alias{print.summary.scrub_FD}
\alias{print.scrub_FD}
\title{Summarize a \code{"scrub_FD"} object}
\usage{
\method{summary}{scrub_FD}(object, ...)

\method{print}{summary.scrub_FD}(x, ...)

\method{print}{scrub_FD}(x, ...)
}
\arguments{
\item{object}{Object of class \code{"scrub_FD"}.}

\item{...}{further arguments passed to or from other methods.}

\item{x}{Object of class \code{"scrub_FD"}.}
}
\value{
A plot of the scrubbing results
}
\description{
Summary method for class \code{"scrub_FD"}
}
