% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_scrub.R
\name{summary.scrub_projection}
\alias{summary.scrub_projection}
\alias{print.summary.scrub_projection}
\alias{print.scrub_projection}
\title{Summarize a \code{"scrub_projection"} object}
\usage{
\method{summary}{scrub_projection}(object, ...)

\method{print}{summary.scrub_projection}(x, ...)

\method{print}{scrub_projection}(x, ...)
}
\arguments{
\item{object}{Object of class \code{"scrub_projection"}.}

\item{...}{further arguments passed to or from other methods.}

\item{x}{Object of class \code{"scrub_projection"}.}
}
\value{
A plot of the scrubbing results
}
\description{
Summary method for class \code{"scrub_projection"}
}
