% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_ACF.R
\name{upper_diag}
\alias{upper_diag}
\title{Yields the upper diagonals of the input matrix.}
\usage{
upper_diag(X, i)
}
\arguments{
\item{X}{The input matrix.}

\item{The}{diagonal index. i=1 is the main diagonal, i=2 is above it and
shorter by 1, etc.}
}
\value{
The diagonal in a numerical vector.
}
\description{
Yields the upper diagonals of the input matrix.
}
