% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_stabilize.R
\name{var_stabilize}
\alias{var_stabilize}
\title{Variance stabilize a timeseries vector}
\usage{
var_stabilize(x, nDCT = 2, lmrob_method = "MM", rescale = TRUE)
}
\arguments{
\item{x}{The timeseries to variance stabilize}

\item{nDCT}{The number of DCT bases to use. Default: \code{4}}

\item{lmrob_method}{The \code{lmrob_method} argument to \code{robustbase::lmrob}.}

\item{rescale}{After variance stabilizing \code{x}, re-center and re-scale
to the original mean and variance? Default: \code{TRUE}.}
}
\value{
The variance stabilized timeseries
}
\description{
Variance detrending implemented by the DCT.
}
