% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{as.matrix2}
\alias{as.matrix2}
\title{Convert CIFTI, NIFTI, or GIFTI file input to \eqn{T} by \eqn{V} matrix}
\usage{
as.matrix2(x, sortSub = FALSE, verbose = FALSE)
}
\arguments{
\item{x}{The object to coerce to a matrix}

\item{sortSub}{Sort subcortex by labels? Default: \code{FALSE}}

\item{verbose}{Print updates? Default: \code{FALSE}}
}
\value{
x as a matrix.
}
\description{
Convert input data to \eqn{T} by \eqn{V} matrix. Reads in a CIFTI, NIFTI,
or GIFTI file. Transposes CIFTI and xifti object input. Masks NIFTI and
nifti object input.
}
\keyword{internal}
