% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choosePCs.R
\name{choose_PCs.kurtosis}
\alias{choose_PCs.kurtosis}
\title{Selects the principle components (PCs) of sufficient kurtosis from a SVD.}
\usage{
choose_PCs.kurtosis(
  svd,
  kurt_quantile = 0.9,
  detrend = TRUE,
  max_keep = NULL,
  min_keep = 1,
  n_sim = 5000
)
}
\arguments{
\item{svd}{An SVD decomposition; i.e. a list containing u, d, and v.}

\item{kurt_quantile}{PCs with kurtosis of at least this quantile are kept.}

\item{detrend}{Should PCs be detrended before measuring kurtosis? Default is
\code{TRUE}. Recommended if observations represent a time series.}

\item{max_keep}{If specified, the total number kept will be at most this
value.}

\item{min_keep}{The total number kept will be at least this
value. The default value is \code{1}.}

\item{n_sim}{The number of simulation data to use for estimating the sampling
distribution of kurtosis. Only used if a new simulation is performed. (If
\eqn{n<1000} and the quantile is 90%, a pre-computed value is used instead.
If \eqn{n>1000}, the theoretical asymptotic distribution is used instead.)}
}
\value{
The original indices of the PCs which were retained, in order of
 decreasing kurtosis.
}
\description{
First, the largest PCs which together explain 90% of the variance are
 retained, and smaller components are removed. Each PC is detrended (this can
 be disabled). The kurtosis cutoff is then the 90% quantile of the sampling
 distribution of kurtosis for Normal data of the same length as the PCs; it
 is estimated by simulation or calculated from the theoretical asymptotic
 distribution if the PCs are long enough. Finally, the total number kept is
 constrained by the \code{max_keep} and \code{min_keep} arguments.
}
