% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choose_PCs.R
\name{choose_PCs.variance}
\alias{choose_PCs.variance}
\title{Selects the principle components of above-average variance from a SVD.}
\usage{
choose_PCs.variance(svd)
}
\arguments{
\item{svd}{An SVD decomposition; i.e. a list containing u, d, and v.}
}
\value{
The original indices of the PCs which were retained, in order of
 decreasing variance (i.e. increasing index).
}
\description{
PCs with above-average variance are retained.
}
