% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choose_PCs.R
\name{choose_PCs.variance}
\alias{choose_PCs.variance}
\title{Selects the components of above-average variance.}
\usage{
choose_PCs.variance(var)
}
\arguments{
\item{var}{The variance of each component, in order. For PCA, this is the 
square of the values on the diagonal of the D matrix.}
}
\value{
The PC indices to retain, in order of decreasing variance (i.e. 
 increasing index).
}
\description{
Components with above-average variance are retained.
}
