% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choosePCs.R
\name{choosePCs_variance}
\alias{choosePCs_variance}
\title{Selects the principle components of sufficient variance from a SVD.}
\usage{
choosePCs_variance(svd, max_keep = NULL, min_keep = NULL)
}
\arguments{
\item{svd}{An SVD decomposition; i.e. a list containing u, d, and v.}

\item{max_keep}{If specified, the total number kept will be at most this
value.}

\item{min_keep}{If specified, the total number kept will be at least this
value.}
}
\value{
The subsetted u matrix with only the chosen columns (PCs).
}
\description{
PCs with above-average variance are retained, but the total number kept is
constrained by the \code{max_keep} and \code{min_keep} arguments.
}
