% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clever.R
\name{clever}
\alias{clever}
\title{Calculates PCA leverage or robust distance and identifies outliers}
\usage{
clever(x, choosePCs = c("mean", "kurtosis"), method = c("leverage",
  "robdist_subset", "robdist"), id_out = TRUE)
}
\arguments{
\item{x}{wide (obs x vars) data matrix of values}

\item{choosePCs}{method to be utilized in choosing which PCs to retain}

\item{method}{method to be utilized in measuring outlyingness}

\item{id_out}{if TRUE (default), will label outliers based on leverage or distance}
}
\value{
A list containing PC scores, number of PCs retained, leverage or distance and (if id.out=TRUE) outlier indicator vectors
}
\description{
Calculates PCA leverage or robust distance and identifies outliers
}
\examples{
n_voxels = 1e4
n_timepoints = 100
x = matrix(rnorm(n_timepoints*n_voxels), ncol = n_voxels)

lev = clever(x)
}
