% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clever.R
\name{clever}
\alias{clever}
\title{Data-driven scrubbing with \code{clever}}
\usage{
clever(
  X,
  measure = "leverage",
  projection = "PCA",
  PESEL = TRUE,
  solve_PC_dirs = FALSE,
  center_X = TRUE,
  scale_X = TRUE,
  detrend_X = 4,
  PCATF_kwargs = NULL,
  kurt_quantile = 0.95,
  get_outliers = TRUE,
  outlier_cutoff = NULL,
  full_PCA = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{X}{Wide numerical data matrix (\eqn{T observations \times V variables}, \eqn{T << V}).
For example, if \code{X} represents an fMRI run, \eqn{T} should be the number
of timepoints and \eqn{V} should be the number of brainordinate vertices/voxels.}

\item{measure}{Choose one of the following scrubbing measures:

 \describe{
   \item{\code{"leverage"}}{(Default) Leverage scrubbing, which is based on
     projecting the data onto directions thought to express outlier 
     information.}
   \item{\code{"DVARS"}}{traditional DVARS}
   \item{\code{"DVARS2"}}{Delta-percent-DVARS and z-score-DVARS (Afyouni and 
     Nichols, 2018)}
 }}

\item{projection}{Only applies if \code{measure} is \code{"leverage"}. Choose
one of the following: \code{"PCA"}, \code{"PCATF"}, or \code{"ICA"}. The
directions of outlyingness will be selected from the components of this
projection.}

\item{PESEL}{Only applies if \code{measure} is \code{"leverage"}. Leverage
 scrubbing selects directions among the largest \eqn{k} components. \eqn{k}
 can be determined by PESEL, a data-driven algorithm (default: \code{TRUE}). 
 Otherwise, \eqn{k} will be the number of above-average-variance PCs 
 (\code{FALSE}). 

 Both methods are based on the principal components (PCs), even if the
 projection is PCATF or ICA.

 For PCA and ICA, the final subset of components will be those in the top
 \eqn{k} which also have scores of high kurtosis (> 95th percentile of the
 kurtosis of data of equal length from a Normal distribution.)}

\item{solve_PC_dirs}{Only applies to the \code{"leverage"} and \code{"robdist"} 
measures. Should the PCA and PCATF principal directions be computed? 
Default: \code{FALSE} (conserves memory). Required to use \code{leverage_images}
for PCA.}

\item{center_X, scale_X}{Center the columns of the data by their medians, and scale the
columns of the data by their median absolute distances (MADs)? Default: \code{TRUE}. Centering is necessary
for detrending and for computing PCA/ICA, so if this is set to \code{FALSE}, 
the input data must already be centered.}

\item{detrend_X}{Detrend the columns of the data using the DCT? Use an integer
 to indicate the number of cosine bases to detrend with (default: \code{4}), or \code{0} 
 to forgo detrending. 

 The data must be centered, either before input or with \code{center_X}.

 Detrending is highly recommended for time-series data, especially if there 
 are many time points or evolving circumstances affecting the data. Additionally,
 for the kurtosis-based projection, trends can induce positive or negative kurtosis,
 contaminating the connection between high kurtosis and outlier presence. 
 
 Detrending should not be used with non-time-series data because the 
 observations are not temporally related.}

\item{PCATF_kwargs}{Options for the \code{"PCATF"} projection. Valid entries 
are: 

\describe{
  \item{K}{Maximum number of PCs to compute. Default: \code{100}. Cannot be set above 100.}
  \item{lambda}{Trend-filtering parameter. Default: \code{0.05}.}
  \item{niter_max}{Maximum number of iterations. Default: \code{1000}.}
  \item{verbose}{Print updates? Default: \code{FALSE}.}
}}

\item{kurt_quantile}{Only applies to PCA and ICA leverage. What cutoff quantile
for kurtosis should be used to select the components? Default: \code{0.95}.}

\item{get_outliers}{Should outliers be flagged based on cutoffs? Default: \code{TRUE}.}

\item{outlier_cutoff}{A number controlling the value of the scrubbing measure
 above which observations are labeled as outliers. If \code{NULL}, use a
 default:

 \describe{
   \item{\code{"leverage"}}{Minimum leverage value, in multiples of the median leverage. Default: \code{4} (will flag leverage scores more than four times the median).}
   \item{\code{"DVARS"}}{Minimum traditional-DVARS value. Default: \code{5}}
   \item{\code{"DVARS2"}}{A length-2 numeric vector representing a dual 
     Delta-percent-DVARS and z-score-DVARS cutoff. Both must be met for a 
     timepoint to be flagged. The Delta-percent-DVARS cutoff should be given
     in percentages; the z-score-DVARS cutoff should be given as a z-score. 
     Or, set this to "Afyouni-Nichols" (default) to require a
     Delta-percent-DVARS of more than 5\% and a z-score-DVARS greater than
     the right-tail 5\% significance level with Bonferroni FWER correction).}
 }}

\item{full_PCA}{Return the full SVD? Default: \code{FALSE} (return
only the components used to compute the measures).}

\item{verbose}{Should occasional updates be printed? Default: \code{FALSE}.}
}
\value{
A \code{"clever"} object, i.e. a list with components
\describe{
 \item{measure}{
   The scrubbing measure. If \code{"DVARS2"}, it will be a \code{data.frame}
   with the first column being DPDVARS and the second being ZDVARS.
 }
 \item{outlier_cutoff}{
   The value of the scrubbing measure used to identify outliers (i.e.
   minimum leverage, traditional DVARS, or DPDVARS and ZDVARS).
 }
 \item{outlier_flags}{
   A length \eqn{T} logical vector. \code{TRUE} values indicate suspected outliers. 
 }
 \item{PCA}{
   If PCA leverage was used, this will be a list with components:
   \describe{
     \item{U}{The \eqn{N x Q} PC score matrix. Only PCs with above-average variance/chosen by PESEL will be included.}
     \item{D}{The standard deviation of each PC. Only PCs with above-average variance/chosen by PESEL will be included.}
     \item{V}{The \eqn{P x Q} PC directions matrix. Included only if \code{solve_PC_dirs}}
     \item{kurt_idx}{The length \code{Q} kurtosis rankings, with 1 indicating the highest-kurtosis PC 
       (among those of above-average variance/chosen by PESEL) and \code{NA} indicating a PC with kurtosis below
       the quantile cutoff. Only included if the "PCA_kurt" projection was used.}
   }
 }
 \item{PCATF}{
   If PCATF leverage was used, this will be a list with components:
   \describe{
     \item{U}{The \eqn{N x Q} PC score matrix. Only PCs with above-average variance/chosen by PESEL will be included.}
     \item{D}{The standard deviation of each PC. Only PCs with above-average variance/chosen by PESEL will be included.}
     \item{V}{The \eqn{P x Q} PC directions matrix. Included only if \code{solve_PC_dirs}}
   }
 }
 \item{ICA}{
   If ICA leverage was used, this will be a list with components:
   \describe{
     \item{S}{The \eqn{P x Q} source signals matrix.} 
     \item{M}{The \eqn{N x Q} mixing matrix. Only ICs with above-average variance/chosen by PESEL will be included.}
     \item{M_dt}{The \eqn{N x Q} detrended mixing matrix. Included only if \code{detrend_X}}
     \item{kurt_idx}{The length \code{Q} kurtosis rankings, with 1 indicating the highest-kurtosis IC 
       (among those of above-average variance/chosen by PESEL) and \code{NA} indicating an IC with kurtosis below
       the quantile cutoff. Only included if the "ICA_kurt" projection was used.}
   }
 }
 \item{call}{The call to this function.}
}
}
\description{
Calculates data-driven scrubbing measures (leverage and DVARS) and identifies
 outliers in high-dimensional data.
}
\examples{
n_voxels = 1e4
n_timepoints = 100
X = matrix(rnorm(n_timepoints*n_voxels), ncol = n_voxels)

clev = clever(X)
}
