% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clever.R
\name{clever}
\alias{clever}
\title{Calculates PCA leverage or robust distance and identifies outliers.}
\usage{
clever(
  X,
  projection_methods = "PCA_kurt",
  outlyingness_methods = "leverage",
  DVARS = TRUE,
  PCs_detrend = TRUE,
  PCATF_kwargs = NULL,
  kurt_quantile = 0.95,
  kurt_detrend = TRUE,
  id_outliers = TRUE,
  lev_cutoff = 4,
  rbd_cutoff = 0.9999,
  lev_images = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{X}{A wide (observations x variables) numerical data matrix.}

\item{projection_methods}{Vector of projection methods to use. Choose at least
 one of the following: \code{"PCA_var"} for PCA + variance, \code{"PCA_kurt"}
 for PCA + kurtosis, and \code{"PCATF"} for PCA Trend Filtering + variance.
 Or, use \code{"all"} to use all methods. Default is \code{c("PCA_kurt")}.

 clever will use all combinations of the requested projection and 
 outlyingness methods that make sense. For example, if  
 \code{projection_methods=c("PCATF", "PCA_var", "PCA_kurt")} and 
 \code{outlyingness_methods=c("leverage", "robdist")} then these five
 combinations will be used: PCATF with leverage, PCA_var with leverage, 
 PCA_var with robdist, PCA_kurt with leverage, and PCA_kurt with robdist. 
 Each method combination will yield its own outlyingness time series.}

\item{outlyingness_methods}{Vector of outlyingness measures to compute. Choose
 at least one of the following: \code{"leverage"} for leverage, or 
 \code{"robdist"} for robust distance. Or, use \code{"all"} to use all methods. 
 Default is \code{c("leverage")}.

 clever will use all combinations of the requested projection and 
 outlyingness methods that make sense. For example, if  
 \code{projection_methods=c("PCATF", "PCA_var", "PCA_kurt")} and 
 \code{outlyingness_methods=c("leverage", "robdist")} then these five
 combinations will be used: PCATF with leverage, PCA_var with leverage, 
 PCA_var with robdist, PCA_kurt with leverage, and PCA_kurt with robdist. Each 
 method combination will yield its own outlyingness time series.}

\item{DVARS}{Should DVARS (Afyouni and Nichols, 2017) be computed too? Default 
is \code{TRUE}.}

\item{PCs_detrend}{Detrend all PCs before computing leverage or robust distance? 
Detrending is recommended for time-series data, especially if there are many
time points or changing circumstances, such as in task-based fMRI. Detrending
should not be used with non-time-series data because the observations are 
not temporally related. Default is \code{TRUE}.}

\item{PCATF_kwargs}{Named list of arguments for PCATF: maximum number of PCs
to compute \code{K} (Default 1000) the trend filtering parameter lambda
(Default \code{0.5}), the number of iterations \code{niter_max} (Default 
\code{1000}), convergence tolerance \code{tol} (Default \code{1e-8}), and option 
to print updates \code{verbose} (Default \code{FALSE}).}

\item{kurt_quantile}{What cutoff quantile for kurtosis should be used? This
argument only applies if \code{"PCA_kurt"} is one of the projection methods 
in \code{"projection_methods"}. Default is \code{0.95}.}

\item{kurt_detrend}{Should the PCs be detrended before measuring kurtosis? 
This argument only applies if \code{"PCA_kurt"} is one of the projection methods 
in \code{"projection_methods"}. Default is \code{TRUE}. Detrending is highly
recommended for time-series data, because trends can induce high kurtosis even
in the absence of outliers. Detrending should not be done with non-time-series
data because the observations are not temporally related.}

\item{id_outliers}{Should the outliers be identified? Default is \code{TRUE}.}

\item{lev_cutoff}{The outlier cutoff value for leverage, as a multiple of the median
leverage. Only used if 
\code{"leverage" \%in\% projection_methods} and \code{id_outliers}. Default is 
\code{4}, or \eqn{4 * median}.}

\item{rbd_cutoff}{The outlier cutoff quantile for MCD distance. Only used if 
\code{"robdist" \%in\% projection_methods} and \code{id_outliers}. Default 
The quantile is computed from the estimated F distribution.}

\item{lev_images}{Should leverage images be computed? If \code{FALSE} memory is
conserved. Default is \code{FALSE}.}

\item{verbose}{Should occasional updates be printed? Default is \code{FALSE}.}
}
\value{
A clever object, i.e. a list with components
\describe{
 \item{params}{A list of all the arguments used.}
 \item{projections}{
   \describe{
     \item{PC_var}{
       \describe{
         \item{indices}{The indices retained from the original SVD decomposition 
           to make the variance-based PC projection.} 
         \item{PCs}{The subsetted SVD decomposition.}  
       }
     }
     \item{PC_kurt}{
       \describe{
         \item{indices}{The indices retained from the original SVD decomposition 
           to make the kurtosis-based PC projection. They are ordered from highest 
           kurtosis to lowest kurtosis.}  
         \item{PCs}{The subsetted SVD decomposition. PCs are ordered in the standard
           way, from highest variance to lowest variance, instead of by kurtosis.}  
       }
     }
     \item{PCATF}{
       \describe{
         \item{indices}{The indices of the trend-filtered PCs used to make the
           projection.}  
         \item{PCs}{The PCATF result.}  
       }
     }
   }
 }
 \item{outlier_measures}{
   \describe{
     \item{PC_var__lev}{The leverage values for the PC_var projection.}
     \item{PC_kurt__lev}{The leverage values for the PC_kurt projection.}
     \item{PCATF__lev}{The leverage values for the PCATF projection.}
     \item{PC_var__rbd}{The robust MCD distance values for the PC_var projection.}
     \item{PC_kurt__rbd}{The robust MCD distance values for the PC_kurt projection.}
     \item{DVARS_DPD}{The Delta percent DVARS values.}
     \item{DVARS_ZD}{The DVARS z-scores.}
   }
 }
 \item{outlier_cutoffs}{
   \describe{
     \item{lev}{The leverage cutoff for outlier detection: \code{lev_cutoff} times
       the median leverage.}
     \item{MCD}{The robust distance cutoff for outlier detection: the 
       \code{rbd_cutoff} quantile of the estimated F distribution.}
     \item{DVARS_DPD}{The Delta percent DVARS cutoff: +/- 5 percent}
     \item{DVARS_ZD}{The DVARS z-score cutoff: the one-sided 5 percent 
       significance level with Bonferroni FWER correction.}
   }
 }
 \item{outlier_flags}{
   \describe{
     \item{PC_var__leverage}{Whether each observation surpasses the outlier cutoff.}
     \item{PC_kurt__leverage}{Whether each observation surpasses the outlier cutoff.}
     \item{PCATF__leverage}{Whether each observation surpasses the outlier cutoff.}
     \item{PC_var__robdist}{Whether each observation surpasses the outlier cutoff.}
     \item{PC_kurt__robdist}{Whether each observation surpasses the outlier cutoff.}
     \item{DVARS_DPD}{Whether each observation surpasses the outlier cutoff.}
     \item{DVARS_ZD}{Whether each observation surpasses the outlier cutoff.}
   }
 }
 \item{MCD_scale}{The scale value for out-of-MCD observations, and NA for
   in-MCD observations. NULL if \code{method} is not robust distance.}
 \item{lev_images}{
   \describe{
     \item{mean}{The average of the PC directions, weighted by the unscaled
       PC scores at each outlying time point (U[i,] * V^T). Row names are
       the corresponding time points.}
     \item{top}{The PC direction with the highest PC score at each outlying
       time point. Row names are the corresponding time points.}
     \item{top_dir}{The index of the PC direction with the highest PC score
       at each outlying time point. Named by timepoint.}
   }
 }
}
}
\description{
Calculates PCA leverage or robust distance and identifies outliers.
}
\examples{
n_voxels = 1e4
n_timepoints = 100
X = matrix(rnorm(n_timepoints*n_voxels), ncol = n_voxels)

clev = clever(X)
}
