% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clever.R
\name{clever}
\alias{clever}
\title{Calculates PCA leverage or robust distance and identifies outliers.}
\usage{
clever(
  X,
  PCA_trend_filtering = TRUE,
  PCA_trend_filtering.kwargs = NULL,
  choose_PCs = c("variance", "kurtosis"),
  kurt_quantile = 0.9,
  kurt_detrend = TRUE,
  method = c("leverage", "robdist_subset", "robdist"),
  id_out = TRUE,
  solve_directions = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{X}{A wide (observations x variables) numerical data matrix.}

\item{PCA_trend_filtering}{Should PCA Trend Filtering (PCATF) be used in place
of regular PCA? Default is \code{TRUE}. If \code{TRUE}, \code{choose_PCs}
must be \code{variance} and \code{method} must be \code{leverage}.}

\item{PCA_trend_filtering.kwargs}{Named list of arguments for \code{PCATF}:
the trend filtering parameter \code{lambda} (Default 0.5), the number of
iterations \code{niter_max} (Default 1000), convergence tolerance \code{tol}
(Default 1e-8), and option to print updates \code{verbose} (Default FALSE).}

\item{choose_PCs}{The criteria for choosing which PCs to retain:
\code{variance} or \code{kurtosis}. Default is \code{variance}. If trend
filtering is being used, this must be \code{variance}.}

\item{kurt_quantile}{The cutoff quantile to use if \code{choose_PCs} is
\code{kurtosis}. Default is 0.9.}

\item{kurt_detrend}{If \code{choose_PCs} is \code{kurtosis}, should PCs be
detrended before measuring kurtosis? Default is \code{TRUE}. Recommended if
observations represent a time series.}

\item{method}{The outlier measurement: \code{leverage}, \code{robdist},
or \code{robdist_subset}. Default is \code{leverage}. If trend filtering
is being used, this must be \code{leverage}.}

\item{id_out}{Should the outliers be identified? Default is \code{TRUE}.}

\item{solve_directions}{Should the principal directions be solved for? These
are needed to display the leverage images for outlying observations. Default
is \code{TRUE}. If the leverage images are not needed, this can be
\code{FALSE} to reduce memory use.}

\item{verbose}{Should occasional updates be printed? Default is \code{FALSE}.}
}
\value{
A clever object, i.e. a list with components
\describe{
	\item{params}{A list of all the arguments used.}
 \item{PCs}{
	\describe{
		\item{indices}{The indices of the selected PCs.}
		\item{svd}{The selected subset of the SVD.}
	}
}
 \item{leverage}{The leverage of each observation. NULL if \code{method} is
	not PCA leverage.}
 \item{robdist}{The robust distance of each observation. NULL if
	\code{method} is not robust distance (subset).}
 \item{inMCD}{Whether each observation is within the MCD subset. NULL if
	\code{method} is not robust distance (subset).}
 \item{outliers}{An n X 3 data.frame indicating if each observation is an
	outlier at each of the three levels.}
}
}
\description{
Calculates PCA leverage or robust distance and identifies outliers.
}
\examples{
n_voxels = 1e4
n_timepoints = 100
X = matrix(rnorm(n_timepoints*n_voxels), ncol = n_voxels)

clev = clever(X)
}
