% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clever_multi.R
\name{clever_multi}
\alias{clever_multi}
\title{Compare multiple leverage measures with \code{clever_multi}}
\usage{
clever_multi(
  X,
  projection = "PCA_kurt",
  nuisance = cbind(1, dct_bases(nrow(X), 4)),
  center = TRUE,
  scale = TRUE,
  comps_mean_dt = FALSE,
  comps_var_dt = FALSE,
  kurt_quantile = 0.99,
  PCATF_kwargs = NULL,
  get_dirs = FALSE,
  full_PCA = FALSE,
  get_outliers = TRUE,
  cutoff = 4,
  verbose = FALSE
)
}
\arguments{
\item{X}{Wide numeric data matrix (\eqn{T observations \times V variables},
\eqn{T << V}). If \code{X} represents an fMRI run, \eqn{T} should be the
number of timepoints and \eqn{V} should be the number of brainordinate
vertices/voxels. The outlyingness of the rows in \code{X} will be measured.}

\item{projection}{Leverage works by projecting the data onto directions likely to
contain outlier information. Choose at least one of the following:

\describe{
\item{\code{"PCA"}}{PCA using the top \eqn{k} PCs.}
\item{\code{"PCA_kurt"}}{PCA using the high-kurtosis PCs among the top \eqn{k}.}
\item{\code{"PCA2"}}{PCA using the top \eqn{k2} PCs.}
\item{\code{"PCA2_kurt"}}{PCA using the high-kurtosis PCs among the top \eqn{k2}.}
\item{\code{"PCATF"}}{PCATF using the top \eqn{k} trend-filtered PCs.}
\item{\code{"PCATF_kurt"}}{PCATF using the high-kurtosis trend-filtered PCs among the top \eqn{k}.}
\item{\code{"PCATF2"}}{PCATF using the top \eqn{k2} trend-filtered PCs.}
\item{\code{"PCATF2_kurt"}}{PCATF using the high-kurtosis trend-filtered PCs among the top \eqn{k2}.}
\item{\code{"ICA"}}{ICA using the top \eqn{k} ICs.}
\item{\code{"ICA_kurt"}}{ICA using the high-kurtosis ICs among the top \eqn{k}.}
\item{\code{"ICA2"}}{ICA using the top \eqn{k2} ICs.}
\item{\code{"ICA2_kurt"}}{ICA using the high-kurtosis ICs among the top \eqn{k2}.}
}

where \eqn{k} is the number of PCs selected by PESEL, and \eqn{k2} is the number
of PCs with above-average variance.

Use \code{"all"} to use all projection methods. Default: \code{"PCA_kurt"}.}

\item{nuisance}{Nuisance signals to regress from each column of \code{X}.
Should be a \eqn{T \times N} numeric matrix where \eqn{N} represents the
number of nuisance signals. Default: a matrix with a constant column
(this is the intercept term in the design matrix) and four DCT bases.
This default nuisance regression will have the effect of detrending the
data by removing low-frequency components. To not perform any nuisance
regression set this argument to \code{NULL}, \code{0}, or \code{FALSE}.

Detrending is highly recommended for time-series data, especially if there
are many time points or evolving circumstances affecting the data. Additionally,
if kurtosis is being used to select the projection directions, trends can
induce positive or negative kurtosis, contaminating the connection between
high kurtosis and outlier presence. Detrending should not be used with
non-time-series data because the observations are not temporally related.

To perform nuisance regression with an intercept, DCT bases, and additional
signals, specify something like \code{cbind(1, dct_bases(nrow(X), 4), more_nuisance)}.}

\item{center}{Center the columns of the data by their medians, and scale the
columns of the data by their median absolute distances (MADs)? Default: \code{TRUE}.
Centering is necessary for computing the projections, so if \code{center} is
\code{FALSE}, the data must already be centered.

Note that centering and scaling occur after nuisance regression, so even if
\code{center} is \code{FALSE}, the data will be centered on the means if
the nuisance regression included an intercept term, as it does by default.}

\item{scale}{Center the columns of the data by their medians, and scale the
columns of the data by their median absolute distances (MADs)? Default: \code{TRUE}.
Centering is necessary for computing the projections, so if \code{center} is
\code{FALSE}, the data must already be centered.

Note that centering and scaling occur after nuisance regression, so even if
\code{center} is \code{FALSE}, the data will be centered on the means if
the nuisance regression included an intercept term, as it does by default.}

\item{comps_mean_dt}{Stabilize the mean and variance of the PCA,
PCATF, and ICA components prior to computing leverage? \code{TRUE},
\code{FALSE} (default), or the number of DCT bases to use for mean and variance detrending
(\code{TRUE} will use 4). Note that these refer to the projection scores and
not the data itself. Also, if scaling but not centering, the components must
be expected to already be centered; otherwise, the results will be invalid.}

\item{comps_var_dt}{Stabilize the mean and variance of the PCA,
PCATF, and ICA components prior to computing leverage? \code{TRUE},
\code{FALSE} (default), or the number of DCT bases to use for mean and variance detrending
(\code{TRUE} will use 4). Note that these refer to the projection scores and
not the data itself. Also, if scaling but not centering, the components must
be expected to already be centered; otherwise, the results will be invalid.}

\item{kurt_quantile}{Only applies to PCA and ICA leverage. What cutoff quantile
for kurtosis should be used to select the components? Default: \code{0.99}.}

\item{PCATF_kwargs}{Arguments to \code{\link{PCATF}} in list form. Valid
entries are:

\describe{
\item{lambda}{Trend-filtering parameter. Default: \code{5}.}
\item{niter_max}{Maximum number of iterations. Default: \code{1000}.}
\item{TOL}{Convergence tolerance parameter. Default: \code{1e-8}.}
\item{verbose}{Print updates? Default: \code{FALSE}.}
}}

\item{get_dirs}{Do the projection directions need to be returned? This is the
\eqn{V} matrix in PCA and \eqn{S} matrix in ICA. The default is \code{FALSE}
to save memory. However, \code{get_dirs==TRUE} is required for \code{\link{lev_images}}.}

\item{full_PCA}{Only applies to the PCA projection. Return the full SVD?
Default: \code{FALSE} (return only the components used to compute the measures).}

\item{get_outliers}{Should outliers be flagged based on \code{cutoff}? Default: \code{TRUE}.}

\item{cutoff}{Median leverage cutoff value. Default: \code{4}.}

\item{verbose}{Should occasional updates be printed? Default: \code{FALSE}.}
}
\value{
A \code{"clever_multi"} object, i.e. a list with components
\describe{
\item{measure}{A \eqn{T \times P} data.frame of numeric leverage values for each of the P projections in \code{projection}.}
\item{outlier_cutoff}{A \eqn{1 \times P} data.frame of numeric outlier cutoff values for each projection (\code{cutoff} times the median leverage).}
\item{outlier_flag}{A \eqn{T \times P} data.frame of logical values where \code{TRUE} indicates suspected outlier presence.}
\item{mask}{
A length \eqn{P} numeric vector corresponding to the data locations in \code{X}. Each value indicates whether the location was masked:
\describe{
\item{1}{The data location was not masked out.}
\item{-1}{The data location was masked out, because it had at least one \code{NA} or \code{NaN} value.}
\item{-2}{The data location was masked out, because it was constant.}
}
}
\item{PCA}{
This will be a list with components:
\describe{
\item{U}{The \eqn{T \times Q} PC score matrix.}
\item{D}{The standard deviation of each PC.}
\item{V}{The \eqn{P \times Q} PC directions matrix. Included only if \code{get_dirs}.}
\item{highkurt}{The length \code{Q} logical vector indicating scores of high kurtosis.}
\item{nPCs_PESEL}{The number of PCs selected by PESEL.}
\item{nPCs_avgvar}{The number of above-average variance PCs.}
}
where \code{Q} is the number of PCs selected by PESEL or of above-average variance (or the greater of the two if both were used).
If PCA was not used, all entries except \code{nPCs_PESEL} and/or \code{nPCs_avgvar} will not be included, depending on which
method(s) was used to select the number of PCs.
}
\item{PCATF}{
If PCATF was used, this will be a list with components:
\describe{
\item{U}{The \eqn{T \times Q} PC score matrix.}
\item{D}{The standard deviation of each PC.}
\item{V}{The \eqn{P \times Q} PC directions matrix. Included only if \code{get_dirs}}
}
}
\item{ICA}{
If ICA was used, this will be a list with components:
\describe{
\item{S}{The \eqn{P \times Q} source signals matrix. Included only if \code{get_dirs}}
\item{M}{The \eqn{T \times Q} mixing matrix.}
\item{highkurt}{The length \code{Q} logical vector indicating mixing scores of high kurtosis.}
}
}
}
}
\description{
Calculates leverage to identify outliers in high-dimensional data.
Can compute multiple kinds of leverage at once.
}
\section{Order of operations}{

The input data undergo a nuisance regression to remove low frequency trends
(see the \code{DCT} argument to control or disable this behavior) and any
other nuisance signals (see the \code{nuisance_too} argument to specify them).
(If \code{DCT==0} and \code{is.null(nuisance_too)}, no nuisance regression
is performed.) Then the data are centered and scaled (see the \code{center}
and \code{scale} arguments to control or disable this behavior). Leverage
is computed afterward.
}

\examples{
n_voxels = 1e4
n_timepoints = 100
X = matrix(rnorm(n_timepoints*n_voxels), ncol = n_voxels)

clev = fMRIscrub:::clever_multi(X)
}
\keyword{internal}
