% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{clever_plot_indiv_panel}
\alias{clever_plot_indiv_panel}
\title{Plots one or several outlyingness measures of the same type.}
\usage{
clever_plot_indiv_panel(meas, cuts, name, MCD_scale = NULL, ...)
}
\arguments{
\item{meas}{A T_ x m data.frame with each column being the time-course for a 
scrubbing method. Column names should identify the method as one of the following:
\code{PCA_var__leverage}, \code{PCA_kurt__leverage}, \code{PCATF__leverage},
\code{PCA_var__robdist}, \code{PCA_kurt__robdist},
\code{DVARS_DPD}, \code{DVARS_ZD}, or \code{FD}.}

\item{cuts}{A 1 x m data.frame with each column being the cutoff for a 
scrubbing method. Column names should be the same as those provided for \code{meas}.}

\item{name}{The name of the type of outlyingness measure being plotted:
\code{Leverage}, \code{RobDist}, \code{RobDist Subset}, \code{FD}, \code{DVARS}.}

\item{MCD_scale}{A T_ x m data.frame that is the scale for out-of-MCD data
and NA for in-MCD data, for each method. Only required and used if the 
measure is robust distance.}

\item{...}{Additional arguments to ggplot: main, sub, xlab, ...}
}
\value{
A ggplot
}
\description{
Plots one or several outlyingness measures of the same type.
}
