% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompCor.R
\name{CompCor}
\alias{CompCor}
\title{CompCor}
\usage{
CompCor(
  vol,
  data_ROI = NULL,
  data = NULL,
  noise_ROI,
  noise_nPC = 5,
  noise_erosion = 0
)
}
\arguments{
\item{vol}{The \eqn{I x J x K x T} volumetric fMRI timeseries, including both
the data voxels (unless \code{data} is provided) and the noise voxels.}

\item{data_ROI}{A logical mask for \code{vol} where \code{TRUE} values
indicate data voxels, from which the noise components will be regressed. Up
to one of \code{data_ROI} or \code{data} should be provided.}

\item{data}{A \eqn{V x T} data matrix from which the noise components
will be regressed. Up to one of \code{data_ROI} or \code{data} should be 
provided.}

\item{noise_ROI}{A list of logical masks for \code{vol} where \code{TRUE} 
values indicate voxels in a certain "noise ROI" (e.g. the WM or CSF). Noise
components will be obtained separately for each noise ROI.}

\item{noise_nPC}{A vector the same length as \code{noise_ROI} indicating
the number of PCs from each noise ROI to use. If this vector is a shorter
length its elements will be recycled. The values can also be between 0 and
1, in which case they will represent the minimum proportion of variance 
explained by the PCs used for each noise ROI. The smallest number of PCs
will be used to achieve this proportion of variance explained. Default:
\code{5}.}

\item{noise_erosion}{The number of layers to erode from each of the 
\code{noise_ROI}. Default: \code{0}.}
}
\value{
A list with entries \code{"data"}, \code{"noise"}, \code{"noise_mask"}
 and \code{"noise_var"}.

 If \code{data_ROI} or \code{data} was provided, the entry \code{"data"} will be a 
 \code{V x T} matrix where each row is a data voxel (if \code{data_ROI} was 
 provided, the voxels are in spatial order; if \code{data} was provided, the 
 voxels are in the same order) time series with each noise PC regressed from
 it. Otherwise, this entry will be \code{NULL}.

 The entry \code{"noise"} is a list of \code{T} \eqn{x} \code{noise_nPC} PC 
 scores, one for each \code{noise_ROI}.
}
\description{
The CompCor algorithm (Behzadi et. al., 2007) 10.1016/j.neuroimage.2007.04.042
}
\details{
The data are centered on each voxel timecourse's median.
}
