% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompCor_HCP.R
\name{CompCor_HCP}
\alias{CompCor_HCP}
\title{Run CompCor on the HCP data}
\usage{
CompCor_HCP(
  nii,
  nii_labels,
  ROI_noise = c("wm_cort", "csf"),
  noise_nPC = 5,
  noise_erosion = NULL,
  timepoints = NULL,
  cii = NULL,
  brainstructures = c("left", "right"),
  center = TRUE,
  scale = TRUE,
  detrend = 0,
  verbose = FALSE
)
}
\arguments{
\item{nii}{\eqn{I \times J \times \K \times T} 
NIFTI object or array (or file path to the NIFTI) which contains
whole-brain data, including the noise ROIs. In the HCP, the corresponding
file is e.g. "../Results/rfMRI_REST1_LR/rfMRI_REST1_LR.nii.gz"}

\item{nii_labels}{\eqn{I \times J \times K}
NIFTI object or array (or file path to the NIFTI) which
contains the corresponding labels to each voxel in \code{nii}. Values should
be according to this table: 
https://surfer.nmr.mgh.harvard.edu/fswiki/FsTutorial/AnatomicalROI/FreeSurferColorLUT .
In the HCP, the corresponding file is "ROIs/Atlas_wmparc.2.nii.gz".}

\item{ROI_noise}{A list of numeric vectors. Each entry should represent labels
 in \code{nii_labels} belonging to a single noise ROI, named by that entry's 
 name. Or, this can be a character vector of at least one of the following: 
 \code{"wm_cort"} (cortical white matter), \code{"wm_cblm"} (cerebellar white
 matter), \code{"csf"} (cerebrospinal fluid). In the latter case, these labels
 will be used:

 \describe{
   \item{wm_cort}{c(3000:4035, 5001, 5002)}
   \item{wm_cblm}{c(7, 46)}
   \item{csf}{c(4, 5, 14, 15, 24, 31, 43, 44, 63, 250, 251, 252, 253, 254, 255))}
 }

 These default ROIs are based on this forum post: 
 https://www.mail-archive.com/hcp-users@humanconnectome.org/msg00931.html

 Default: \code{c("wm_cort", "csf")}}

\item{noise_nPC}{The number of principal components to compute for each noise
 ROI. Alternatively, values between 0 and 1, in which case they will 
 represent the minimum proportion of variance explained by the PCs used for
 each noise ROI. The smallest number of PCs will be used to achieve this 
 proportion of variance explained. 

 Should be a list or numeric vector with the same length as \code{ROI_noise}. 
 It will be matched to each ROI based on the name of each entry, or if the 
 names are missing, the order of entries. If it is an unnamed vector, its
 elements will be recycled. Default: \code{5} (compute the top 5 PCs for 
 each noise ROI).}

\item{noise_erosion}{The number of voxel layers to erode the noise ROIs by. 
Should be a list or numeric vector with the same length as \code{ROI_noise}. 
It will be matched to each ROI based on the name of each entry, or if the 
names are missing, the order of entries. If it is an unnamed vector, its 
elements will be recycled. Default: \code{NULL}, which will use a value of
0 (do not erode the noise ROIs).}

\item{timepoints}{A numeric vector indicating the timepoints to compute 
CompCor for, or \code{NULL} (default) to use all timepoints. (Indexing begins
with 1, so the first timepoint has index 1 and the last has the same index
as the length of the scan.)}

\item{cii}{\code{"xifti"} (or file path to the CIFTI) from which the noise
ROI components will be regressed. In the HCP, the corresponding file is e.g.
"../Results/rfMRI_REST1_LR/rfMRI_REST1_LR_Atlas_MSMAll.dtseries.nii"}

\item{brainstructures}{Choose among "left", "right", and "subcortical".
Default: \code{c("left", "right")} (cortical data only)}

\item{center, scale, detrend}{Center, scale, and detrend data columns? Will
affect both the NIFTI noise ROIs and the CIFTI greyordinate data. Centering
and scaling is \code{TRUE} or \code{FALSE} where as detrending should be
indicated by the number of DCT bases to regress (0 to not detrend). By default,
the data is centered and scaled but not detrended.}

\item{verbose}{Should occasional updates be printed? Default: \code{FALSE}.}
}
\description{
Wrapper to \code{\link{clever_multi}} for computing CompCor (and other outlyingness
 measures) on HCP data. The whole-brain NIFTI is used to obtain the noise
 ROIs, which are regressed from the greyordinate data in the CIFTI.
}
