% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Dat1}
\alias{Dat1}
\title{First Example Time Series from the ABIDE}
\format{
A numeric matrix of 193 time points by 4675 voxels
}
\usage{
Dat1
}
\description{
A sagittal slice from the fMRI time series for subject 0050048. The scan was
obtained at the University of Pittsburgh School of Medicine. The scan has
been pre-processed with slice time correction, rigid body realignment
estimation, spatial normalization to MNI space, and linear detrending.
Subject 0050048 was a typically-developing 11-year-old male. The scan has
many artifacts. A mask was applied to vectorize the spatial
dimensions.
}
\details{
Source: http://fcon_1000.projects.nitrc.org/indi/abide/abide_I.html
}
\section{References}{

\itemize{
\item{1.Di Martino, A. et al. The autism brain imaging data exchange: towards a large-scale evaluation of the intrinsic brain architecture in autism. Mol Psychiatry 19, 659–667 (2014).}
}
}

\keyword{datasets}
