% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DVARS.R
\name{DVARS}
\alias{DVARS}
\title{Computes the DSE decomposition and DVARS-related statistics.
Citation: Insight and inference for DVARS (Afyouni and Nichols, 2018)}
\usage{
DVARS(
  X,
  normalize = FALSE,
  cutoff_DVARS = NULL,
  cutoff_DPD = 5,
  cutoff_ZD = qnorm(1 - 0.05/nrow(as.matrix(X))),
  verbose = FALSE
)
}
\arguments{
\item{cutoff_DVARS, cutoff_DPD, cutoff_ZD}{Numeric outlier cutoffs. Timepoints
exceeding these cutoffs will be flagged as outliers.}

\item{verbose}{Should occasional updates be printed? Default is \code{FALSE}.}
}
\description{
Differences from implementation at github.com/asoroosh/DVARS:
\itemize{
 \item The matrix is transposed.
 \item We center and scale the matrix differently (see \code{\link{scale_med}})
 \item We set all zero-variance voxels to zero during centering & scaling. This means that when we remove constant 0 or NA voxels, constant non-zero voxels are also removed.
 \item We use a tolerance of \eqn{1e-8} to detect non-zero voxels.
}
}
\details{
\code{Y <- Y/100; DVARS(t(Y - apply(Y, 1, mean)))} where \code{Y} is the 
 \eqn{V \times T} data matrix.
}
