% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DVARS.R
\name{DVARS}
\alias{DVARS}
\title{Computes the DSE decomposition and DVARS-related statistics.
Citation: Insight and inference for DVARS (Afyouni and Nichols, 2018)}
\usage{
DVARS(X, normalize = FALSE, norm_I = 100, verbose = FALSE)
}
\arguments{
\item{X}{a T x N numeric matrix representing an fMRI run.}

\item{normalize}{Normalize the data as proposed in the original paper? Default is 
\code{FALSE}.}

\item{norm_I}{The value to scale to. Default is \code{100}, as in the original
paper.}

\item{verbose}{Should occasional updates be printed? Default is \code{FALSE}.}
}
\description{
Differences from implementation at github.com/asoroosh/DVARS:
\itemize{
 \item The matrix is transposed.
 \item We center and scale the matrix differently (see \code{\link{scale_med}})
 \item We set all zero-variance voxels to zero during centering & scaling. This means that when we remove constant 0 or NA voxels, constant non-zero voxels are also removed.
 \item We use a tolerance of \eqn{1e-8} to detect non-zero voxels.
}
}
