% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_ACF.R
\name{est_ACF}
\alias{est_ACF}
\title{Estimates the ACF of columns in X.}
\usage{
est_ACF(X, detrend = TRUE, cor_method = "spearman",
  diag_center_method = median, beta = 0.1,
  ACF_est_method = "tapering", k = NULL)
}
\arguments{
\item{X}{A txp matrix representing p length-t time series.}

\item{detrend}{Should the time series be detrended? Default TRUE.}

\item{cor_method}{The \code{method} parameter for \code{stats::cor} to use
for ACF estimation. Default is 'spearman'.}

\item{diag_center_method}{How ACF estimates for each column should be
aggregated. Default is \code{stats::median}.}

\item{beta}{The beta value for the tapering/banding estimator. See
\code{k_choice()}.}

\item{ACF_est_method}{Whether to use the 'tapering' or 'banding' estimator
from Cai, Ren and Zhou, 2013.}

\item{The}{k value for the tapering/banding estimator. Overrides \code{beta}.
See \code{k_choice()}.}
}
\value{
The ACF as a numeric vector.
}
\description{
Estimates the ACF of columns in X.
}
