% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{est_trend}
\alias{est_trend}
\title{Estimates the trend of \code{ts} using a robust discrete cosine transform.}
\usage{
est_trend(ts, robust = TRUE)
}
\arguments{
\item{ts}{A numeric vector to detrend.}

\item{robust}{Should a robust linear model be used? Default FALSE.}
}
\value{
The estimated trend.
}
\description{
Estimates the trend of \code{ts} using a robust discrete cosine transform.
}
