% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FD.R
\name{FD}
\alias{FD}
\title{Framewise Displacement}
\usage{
FD(
  X,
  trans_units = c("mm", "cm", "in"),
  rot_units = c("rad", "deg", "mm", "cm", "in"),
  brain_radius = NULL,
  detrend = FALSE
)
}
\arguments{
\item{X}{An \eqn{N \times 6} matrix in which the first three columns represent the
 translational RPs (\code{trans_units}), and the second three columns represent
 the rotational RPs (\code{rot_units}). If \code{rot_units} measures an angle,
 it will be converted to \code{trans_units} by measuring displacement on a 
 sphere of radius \code{brain_radius} \code{trans_units}.

 Alternatively, this can be the file path to an \eqn{N \times 6} matrix which can be
 read with \code{\link[utils]{read.table}} (fields separated by white-space; no
 header).}

\item{trans_units}{\code{"mm"} for millimeters (default), \code{"cm"} 
for centimeters, or \code{"in"} for inches.}

\item{rot_units}{\code{"rad"} for radians (default), \code{"deg"} for degrees,
or one of the \code{trans_units} options.}

\item{brain_radius}{If \code{rot_units} measures an angle, the rotational RPs
 are transformed to a spatial measurement representing the displacement on a 
 sphere of radius \code{brain_radius} \code{trans_units}.

 If \code{brain_radius} is \code{NULL} (default), its value will be set to 
 (the equivalent of) 50 mm.}

\item{detrend}{Detrend each RP with \code{\link{est_trend}} before computing FD?
Default: \code{FALSE}.}
}
\value{
A length \eqn{N} vector of FD values in \code{trans_units}.
}
\description{
Calculates Framewise Displacement (FD)
}
\details{
The FD formula is taken from (Power, 2012):

 \eqn{FD_i = ∣\Delta x_i∣ + ∣\Delta y_i∣ + ∣\Delta z_i + ∣\Delta \alpha_i∣ + ∣\Delta \beta_i∣ + ∣\Delta \gamma_i∣}, 
 where \eqn{i} is the timepoint; \eqn{x}, \eqn{y} and \eqn{z} are the 
 translational realignment parameters (RPs);
 \eqn{\alpha}, \eqn{\beta} and \eqn{\gamma} are the rotational RPs;
 and \eqn{\Delta x_i = x_{i-1} − x_i} (and similarly for the other RPs).

 Citation: 10.1016/j.neuroimage.2011.10.018
}
