% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leverage_images.R
\name{get_leverage_images}
\alias{get_leverage_images}
\title{Calculate the leverage images}
\usage{
get_leverage_images(clev, timepoints = NULL, const_mask = NULL)
}
\arguments{
\item{clev}{A "clever" object.}

\item{timepoints}{The times for which to compute leverage images (rows of U).}

\item{const_mask}{Mask that is TRUE where voxels were removed.}
}
\value{
A list of three: the mean leverage image for each outlier meeting
 the thresold, the top leverage image for each outlier, and the indices of
 the top leverage images.
}
\description{
Calculate the leverage images
}
