% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{leverage_images}
\alias{leverage_images}
\title{Calculate the leverage images for each outlier that meets the
 \code{outlier_level} threshold, with 3 (default) being the highest/strictest
and 1 being the lowest.}
\usage{
leverage_images(x, outlier_level = 3)
}
\arguments{
\item{x}{A clever object.}

\item{outlier_level}{The outlier threshold for the images, with 3 (default)
being the highest/strictest. If no outliers at or above this threshold
exist, no images will be made.}
}
\value{
A list of three: the mean leverage images for each outlier meeting
 the thresold, the top leverage images, and the indices of the top leverage
 images.
}
\description{
Calculate the leverage images for each outlier that meets the
 \code{outlier_level} threshold, with 3 (default) being the highest/strictest
and 1 being the lowest.
}
