% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{leverage_images}
\alias{leverage_images}
\title{Calculate the leverage images for each outlier that meets the
 \code{outlier_level} threshold, with 3 (default) being the highest/strictest
and 1 being the lowest.}
\usage{
leverage_images(svd, timepoints)
}
\arguments{
\item{svd}{A singular value decomposition (list with u, v, and d).}

\item{timepoints}{The times for which to compute leverage images (rows of U).}
}
\value{
A list of three: the mean leverage image for each outlier meeting
 the thresold, the top leverage image for each outlier, and the indices of
the top leverage images.
}
\description{
Calculate the leverage images for each outlier that meets the
 \code{outlier_level} threshold, with 3 (default) being the highest/strictest
and 1 being the lowest.
}
