% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R, R/visualize.R
\name{Matrix_to_VolumeTimeSeries}
\alias{Matrix_to_VolumeTimeSeries}
\title{Converts a vectorized matrix back to a volume time series.}
\usage{
Matrix_to_VolumeTimeSeries(mat, mask, sliced.dim = NA)

Matrix_to_VolumeTimeSeries(mat, mask, sliced.dim = NA)
}
\arguments{
\item{mat}{A matrix whose rows are observations at different times, and
columns are pixels/voxels.}

\item{mask}{A corresponding binary mask, with 1's representing regions
within the area of interest and 0's representing regions to mask out.}

\item{sliced.dim}{If the mask is 2D, which dimension does it represent?
Will default to the 3rd dimension (axial).}
}
\value{
A 4D array representing the volume time series. Time is on the 4th
 dimension.
}
\description{
Converts a vectorized matrix back to a volume time series.

Applies a 2D/3D mask to a matrix to get a 3D/4D volume time series.
}
