% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca_leverage.R
\name{pca_leverage}
\alias{pca_leverage}
\title{Calculate the Leverage for a Principal Components Analysis}
\usage{
pca_leverage(x)
}
\arguments{
\item{x}{data matrix of values, passed to \code{\link{fast.prcomp}}}
}
\value{
A list of components and the leverage
}
\description{
Calculate the Leverage for a Principal Components Analysis
}
\examples{
x = matrix(rnorm(100*3), ncol = 3)
lev = pca_leverage(x)
}
