% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCATF.R
\name{PCATF}
\alias{PCATF}
\title{PCA Trend Filtering. From: https://github.com/Lei-D/PCATF}
\usage{
PCATF(
  X,
  X.svd = NULL,
  solve_directions = TRUE,
  K = NULL,
  max_K = 100,
  lambda = 0.5,
  niter_max = 1000,
  tol = 1e-08,
  verbose = FALSE
)
}
\arguments{
\item{X}{A numerical data matrix (observations x variables).}

\item{X.svd}{(Optional) The svd decomposition of X. Save time by providing
this argument if the svd has already been computed. Default NULL.}

\item{solve_directions}{Should the principal directions be solved for? These
will be needed to display the leverage images for outlying observations.}

\item{K}{(Optional) The number of PCs to solve for. If not provided, the
number of PCs will be the amount of regular PCs with variance above
the mean, up to 100 PCs.}

\item{lambda}{The trend filtering parameter; roughly, the filtering intensity.
Default is .5 . Can be NULL (lets algorithm decide).}

\item{niter_max}{The number of iterations to use for approximating the PC.}

\item{tol}{The maximum 2-norm between iterations to accept as convergence.}

\item{verbose}{Print statements about convergence?}
}
\value{
SVD The trend-filtered SVD decomposition of X (list with u, d, v).
}
\description{
PCA Trend Filtering. From: https://github.com/Lei-D/PCATF
}
