% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_levdist.R
\name{PCleverage}
\alias{PCleverage}
\title{Computes PCA leverage.}
\usage{
PCleverage(U)
}
\arguments{
\item{U}{An n x Q matrix of PC scores.}
}
\value{
A vector of length n with the leverage of each observation.
}
\description{
Computes the leverage of each observation, based on its PC scores.
}
