% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.clever}
\alias{plot.clever}
\title{Plot \code{"clever"}}
\usage{
\method{plot}{clever}(x, measures = "all", title = NULL, ...)
}
\arguments{
\item{x}{The \code{"clever"} object.}

\item{measures}{"all" to plot each measure (default), or a character vector 
of desired measures.}

\item{title}{(Optional) If provided, will add a title to the plot.}

\item{...}{Additional arguments to ggplot: main, sub, xlab, ...}
}
\value{
A ggplot
}
\description{
Plots the outlyingness measures from a \code{"clever"} result. Can support 
 multiple panels of different outlyingness measures, but by default, 
 it will plot only the first measures.
}
