% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{plot.clever}
\alias{plot.clever}
\title{Visualizes the outlier distribution of a clever object}
\usage{
\method{plot}{clever}(clever, log_measure = FALSE)
}
\arguments{
\item{clever}{a clever object}

\item{log_measure}{if the measure (on y-axis), i.e. mean or kurtosis, should be log transformed (base 10)}
}
\description{
Visualizes the outlier distribution of a clever object
}
