% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{plot.clever}
\alias{plot.clever}
\title{Visualizes the outlier distribution of a clever object.}
\usage{
\method{plot}{clever}(clev, ...)
}
\arguments{
\item{clev}{A clever object.}
}
\value{
The clever ggplot.
}
\description{
Prints a dot plot of the observations (x-axis) against their outlyingness (y-axis), i.e.
 their leverage or robust distance.
}
\details{
Cutoffs for each outlier level are marked by horizontal dashed lines. Outliers are 
 highlighted by vertical lines which extend down to the x-axis; they are colored yellow, 
 orange and red in order of increasing outlyingness.

If the outlyingness measure is robust distance, observations within the MCD are plotted
 separately from those outside the MCD. Also, the y-axes will be log10-scaled.
}
