% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.clever_multi}
\alias{plot.clever_multi}
\title{Plot \code{"clever_multi"}}
\usage{
\method{plot}{clever_multi}(x, title = NULL, ...)
}
\arguments{
\item{x}{The \code{"clever_multi"} object.}

\item{title}{(Optional) If provided, will add a title to the plot.}

\item{...}{Additional arguments to ggplot, e.g. \code{main}, \code{sub},
\code{xlab}, \code{ylab}, \code{legend.position}}
}
\value{
A ggplot
}
\description{
Plot \code{"clever_multi"}
}
\keyword{internal}
