% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{scale_med}
\alias{scale_med}
\title{Centers and scales a matrix robustly for the purpose of covariance estimation.}
\usage{
scale_med(mat)
}
\arguments{
\item{mat}{A numerical matrix.}
}
\value{
The input matrix centered and scaled.
}
\description{
Centers each column on its median, and scales each column by its median absolute deviation (MAD).
 If any column MAD is zero, its values become zero and a warning is raised. If all MADs are 
 zero, an error is raised.
}
