% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SHASH.R
\name{SHASH_out}
\alias{SHASH_out}
\title{Robust outlier detection based on SHASH distribution}
\usage{
SHASH_out(x, maxit = 100)
}
\arguments{
\item{x}{The numeric vector in which to detect outliers.}

\item{maxit}{The maximum number of iterations. Default: \code{10}.}
}
\value{
A \code{"SHASH_out"} object, i.e. a list with components
\describe{
\item{out_idx}{Indices of the detected outliers.}
\item{indx_iters}{TRUE for the detected outliers for each itertation.}
\item{last_iter}{Last iteration number.}
\item{converged}{Logical indicating whether the convergence criteria was satisfied or not.}
}
}
\description{
A robust outlier detection based on modeling the data as coming from a SHASH
distribution.
}
